% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\docType{class}
\name{adaptor-class}
\alias{adaptor-class}
\title{adaptor}
\value{
An adaptor object.
}
\description{
The adaptor class is an externalptr object that points to related C++ objects.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{An externalptr object.}
}}

