% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptorCreatingFunctions.R
\name{adaptorUsingFixedParam}
\alias{adaptorUsingFixedParam}
\title{Create an adaptor using fixed values}
\usage{
adaptorUsingFixedParam(fixedParam)
}
\arguments{
\item{fixedParam}{A numeric vector indicating the fixed parameter values.}
}
\value{
An adaptor object, which is an S4 object of class "adaptor".
}
\description{
Creates an adaptor using a specified numeric vector of fixed parameter values.
This is especially useful for setting known values of parameters that will not change during estimation.
For instance, one might set the 'n' parameter to 1 for a Bernoulli distribution (which is a special case of the binomial distribution).
}
\examples{
\dontrun{
  # Define a fixed parameter for a Bernoulli trial
  fixedValue <- 1
  # Create an adaptor using the fixed parameter
  myAdaptor <- adaptorUsingFixedParam(fixedParam = fixedValue)
  class(myAdaptor) # should return "adaptor"

  # How to use the adaptor in practice:
  # Here, we set the 'n' parameter of a Binomial distribution as a fixed value, and
  # the 'prob' parameter is set using an index pointing adaptor
  cgf = BinomialModelCGF(n = myAdaptor, prob = adaptor.using.indices(1))
  # Evaluate K using the created CGF object
  K(tvec = 0, parameter.vector = 0.5, baseCGF = cgf)
}

}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}
}
