% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptorCreatingFunctions.R
\name{adaptorUsingIndices}
\alias{adaptorUsingIndices}
\title{Create an adaptor using indices}
\usage{
adaptorUsingIndices(indices)
}
\arguments{
\item{indices}{A numeric vector of positive integers representing the subsetting indices.}
}
\value{
An adaptor object, which is an S4 object of class "adaptor".
}
\description{
Creates an adaptor using a specified numeric vector of positive integer indices.
These indices are intended for subsetting operations.
This function belongs to a suite of functions designed to generate adaptor objects.
}
\examples{
\dontrun{
  # Create an adaptor using an index pointing to the third position in the parameter vector
  myAdaptor <- adaptorUsingIndices(indices = 3)
  class(myAdaptor) # should return "adaptor"

  # Using the adaptors
  # Here, the 'n' parameter is set to a fixed value of 1 using (\code{\link{adaptorUsingFixedParam}})
  # and the 'prob' parameter will be fetched from the third position of the
  # `parameter.vector` due to the index specified in `myAdaptor`.
  cgf = BinomialModelCGF(n = adaptorUsingFixedParam(1), prob = myAdaptor)

  # The parameter.vector has the value 0.5 at the third position, which will be used as the 'prob' parameter
  K(tvec = 0, parameter.vector = c(345, 9, 0.5), baseCGF = cgf)
}

}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingFixedParam}}
}
