% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptorCreatingFunctions.R
\name{adaptorUsingRFunctions}
\alias{adaptorUsingRFunctions}
\title{Create an adaptor Using R Functions}
\usage{
adaptorUsingRFunctions(h, grad_h)
}
\arguments{
\item{h}{A function that computes h(theta). The function should accept a numeric vector as input and return a numeric vector.}

\item{grad_h}{A function that computes the gradient of h(theta) with respect to theta. It should accept a numeric vector as input and return a matrix with dimensions length(h(theta)) rows and length(theta) columns}
}
\value{
An adaptor object, which is an S4 object of class "adaptor".
}
\description{
Creates an adaptor object using provided R functions. This is useful when specific transformations
of the parameter space are desired, or for implementing a derivative of the function with respect to parameters.
}
\examples{
\dontrun{
  # Example functions h and grad.h
  h = function(theta) c(theta[1], 5, 3, theta[2], theta[3]) # theta is a numeric vector of length 3
  grad.h = function(theta) matrix(c(1,0,0,0,0, 0,0,0,1,0, 0,0,0,0,1), ncol = 3)

  myAdaptor1 = adaptorUsingRFunctions(h, grad.h)
  class(myAdaptor1) # should return "adaptor"

  # Another set of functions
  f = function(theta) c(theta^2, 2*theta, theta)
  grad.f = function(theta) {
     matrix(c(2*theta, 2, 1), ncol = 1)
  }
  myAdaptor2 = adaptorUsingRFunctions(f, grad.f)
  class(myAdaptor2) # should return "adaptor"
}

}
\seealso{
\code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
