% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-BinomialCGF.R
\docType{data}
\name{BinomialCGF}
\alias{BinomialCGF}
\title{Binomial CGF Object}
\format{
An object of class \code{CGF} (R6), with standard methods:
\code{K}, \code{K1}, \code{K2}, \code{K3operator}, etc.
}
\usage{
BinomialCGF
}
\description{
Ready-to-use CGF for Binomial(n, p). Accepts packed params \code{c(n_vec, p_vec)}.
If \code{length(tvec)} is a multiple of \code{length(n_vec)}, evaluation proceeds (iidReps = "any").
}
\details{
By default, \code{BinomialCGF} supports vectorized evaluation for i.i.d. replicates.
}
\examples{
# Expected value of X ~ Binomial(10, 0.3)
BinomialCGF$K1(0, c(10, 0.3))

}
\keyword{datasets}
