% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-BinomialCGF.R
\name{BinomialModelCGF}
\alias{BinomialModelCGF}
\title{Create a Parametric Binomial CGF Object}
\usage{
BinomialModelCGF(n, prob, iidReps = "any", ...)
}
\arguments{
\item{n}{A function or \code{adaptor} mapping \code{theta} -> scalar or vector of trial counts.}

\item{prob}{A function or \code{adaptor} mapping \code{theta} -> scalar or vector of probabilities.}

\item{iidReps}{Either \code{"any"} or a positive integer. Default \code{"any"}.}

\item{...}{Passed through to the CGF creation (e.g., to override operators).}
}
\value{
A \code{CGF} object.
}
\description{
Binomial with parameters \code{n(theta)} and \code{prob(theta)}. If \code{iidReps = "any"} (default),
\code{length(tvec)} must be a multiple of \code{length(n(theta))}. If \code{iidReps = m}, then
\code{length(tvec)} must be \code{m * length(n(theta))}. It supports i.i.d. replication (via \code{iidReps}) and non-identical usage by
causing \code{n(theta)} and \code{prob(theta)} to return vectors.
}
\examples{
# i.i.d. example: n=10, p=0.3
n_fn <- function(th) th[1]
p_fn <- function(th) th[2]
my_cgf <- BinomialModelCGF(n_fn, p_fn, iidReps = 1)
# If iidReps="any", length(tvec) determines the number of i.i.d. replicates
my_cgf$K1(0, c(10, 0.3))

# non-identical example: n=c(5,10), p=c(0.2,0.7)
# param ==> c(5,10, 0.2, 0.7)
n_adapt <- function(th) th[1:2] # OR n_adapt <- adaptor(indices = 1:2)
p_adapt <- function(th) th[3:4] # OR p_adapt <- adaptor(indices = 3:4)
my_cgf2 <- BinomialModelCGF(n_adapt, p_adapt, iidReps="any")
# default iidReps="any".
# If iidReps=m, then m-i.i.d. blocks are expected => length(tvec)
# must be a multiple of m
# e.g. tvec= c(0,0) => length=2 => we have 2 binomial distributions: (5,0.2) and (10,0.7)
my_cgf2$K1(c(0,0), c(5,10,0.2,0.7))

}
