% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-class.R
\name{CGF}
\alias{CGF}
\title{CGF Constructor}
\usage{
CGF(ptr)
}
\arguments{
\item{ptr}{An externalptr object. It should not be an empty externalptr object.}
}
\value{
An object of class "CGF".
}
\description{
This function constructs an object of class "CGF" using a valid externalptr object.
}
\examples{
\dontrun{
  # assuming ptr is a valid externalptr object
  cgf <- CGF(ptr)
}
}
