% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-funcT.R
\name{compute.funcT}
\alias{compute.funcT}
\title{Compute the correction term for the saddlepoint approximation to the log-likelihood}
\usage{
compute.funcT(
  theta,
  observed.data,
  cgf,
  tvec.hat = NULL,
  gradient = FALSE,
  hessian = FALSE,
  zeroth.order = FALSE,
  ...
)
}
\arguments{
\item{theta}{Numeric vector of parameters.}

\item{observed.data}{Numeric vector of observed data.}

\item{cgf}{A CGF object.}

\item{tvec.hat}{(Optional) Numeric vector. If supplied, \code{tvec} is taken directly
as this vector (the saddlepoint \eqn{\hat{t}}). Otherwise, we solve for it or try
\code{analytic_tvec_hat}.}

\item{gradient}{Logical. If \code{TRUE}, the gradient wrt \code{theta} is returned.}

\item{hessian}{Logical. If \code{TRUE}, the Hessian wrt \code{theta} is returned.}

\item{zeroth.order}{Logical. Default is \code{FALSE}.}

\item{...}{Optional arguments to be passed to \code{\link{saddlepoint.solve}}.}
}
\value{
A named list with elements:
\describe{
\item{val}{Numeric vector of the correction term.}
\item{gradient}{The gradient at \code{theta}, if \code{gradient=TRUE}.}
\item{hessian}{The Hessian at \code{theta}, if \code{hessian=TRUE}.}
}
}
\description{
This function calculates the correction term for the saddlepoint approximation
to the log-likelihood. By default, it computes that of the \strong{first-order} approximation,
If \code{zeroth.order=TRUE}, it instead computes the correction term for the \strong{zeroth-order} approximation.
}
\seealso{
\code{\link{find.saddlepoint.MLE}}
}
