% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-functions.R
\name{create_saddlepoint.ineq.constraint_function}
\alias{create_saddlepoint.ineq.constraint_function}
\title{Create saddlepoint (CGF-based) inequality constraint function}
\usage{
create_saddlepoint.ineq.constraint_function(tvec, theta, model.cgf)
}
\arguments{
\item{tvec}{A numeric vector.}

\item{theta}{A numeric vector of model parameters.}

\item{model.cgf}{An object of class 'CGF'.}
}
\value{
A function that accepts a single vector argument 'a'. When \code{a = c(tvec, theta)} is passed, the function yields a list in the form \code{list(constraints = , jacobian = )}, where 'constraints' are the calculated inequalities and 'jacobian' is the gradient of the constraints with respect to both \code{tvec} and \code{theta}.
}
\description{
This function constructs an inequality constraint function based on the model CGF.
}
\keyword{internal}
