% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-CGF_fromMultinomialFamily.R
\name{createMultinomialFamilyCGF}
\alias{createMultinomialFamilyCGF}
\title{Create a MultinomialFamilyCGF object}
\usage{
createMultinomialFamilyCGF(
  K_func = NULL,
  K1_func = NULL,
  K2_func = NULL,
  K3operator_func = NULL,
  K4operator_func = NULL,
  ineq_constraint_func = NULL,
  analytic_tvec_hat_func = NULL,
  tilting_exponent_func = NULL,
  neg_ll_func = NULL,
  func_T_func = NULL,
  K4operatorAABB_func = NULL,
  K3K3operatorAABBCC_func = NULL,
  K3K3operatorABCABC_func = NULL,
  K4operatorAABB_factored_func = NULL,
  K3K3operatorAABBCC_factored_func = NULL,
  K3K3operatorABCABC_factored_func = NULL,
  K2operator_func = NULL,
  K2operatorAK2AT_func = NULL,
  op_name = "UnnamedOperation",
  ...
)
}
\arguments{
\item{K_func, K1_func, K2_func, K3operator_func, K4operator_func}{The five required methods; if \code{NULL},
use the multinomial default in the child class.}

\item{ineq_constraint_func}{Optional function (default \code{NULL} => no constraints).}

\item{analytic_tvec_hat_func}{Optional function.}

\item{tilting_exponent_func, neg_ll_func, func_T_func}{Optional overrides.}

\item{K4operatorAABB_func, K3K3operatorAABBCC_func, K3K3operatorABCABC_func}{Optional overrides.}

\item{K4operatorAABB_factored_func, K3K3operatorAABBCC_factored_func, K3K3operatorABCABC_factored_func}{Optional overrides.}

\item{K2operator_func, K2operatorAK2AT_func}{Optional overrides.}

\item{op_name}{Character string operation label. Default "UnnamedOperation".}

\item{...}{Additional arguments for future use or to pass to the base CGF's \code{initialize}.}
}
\value{
A CGF object.
}
\description{
This factory function instantiates a CGF object related to the multinomial distribution.
It accepts arguments for all required and optional CGF methods.
If a required method (K_func, K1_func, etc.) is \code{NULL}, the class's built-in
multinomial default is used. For optional methods that are \code{NULL}, we rely on
the base CGF defaults.
}
