% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{DGP}
\alias{DGP}
\title{Data Generating Process}
\usage{
DGP(n, maf = 0.25, pve_g = 0, pve_x = 0.2)
}
\arguments{
\item{n}{Sample size.}

\item{maf}{Genotype minor allele frequency.}

\item{pve_g}{Proportion of variation explained by genotype.}

\item{pve_x}{Proportion of variation explained by covariates.}
}
\value{
Data.frame.
}
\description{
Simulates example data.
}
