% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-rtmb-getValue.R
\name{.rtmb_value_real}
\alias{.rtmb_value_real}
\title{Extract numeric values from RTMB AD objects (without RTMB:::)}
\usage{
.rtmb_value_real(x)
}
\description{
Extract numeric values from RTMB AD objects (without RTMB:::)
}
\keyword{internal}
