% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecgf.R
\name{ECGFR}
\alias{ECGFR}
\title{Empirical CGF of R}
\usage{
ECGFR(mu, y)
}
\arguments{
\item{mu}{Mean of Y under the null.}

\item{y}{Phenotype.}
}
\value{
Function, ECGF(t).
}
\description{
Calculate the empirical cumulant generating function of the residual.
}
