% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecgf.R
\name{ECGFS}
\alias{ECGFS}
\title{Empirical CGF of S}
\usage{
ECGFS(g, mu, y)
}
\arguments{
\item{g}{Genotype.}

\item{mu}{Mean of Y under the null.}

\item{y}{Phenotype.}
}
\value{
Function, ECGF(t).
}
\description{
Calculate the empirical cumulant generating function of the score.
}
