% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ECGFScpp}
\alias{ECGFScpp}
\title{Score ECGF}
\usage{
ECGFScpp(g, mu, t, y)
}
\arguments{
\item{g}{Genotype.}

\item{mu}{Mean of Y under the null.}

\item{t}{Evaluation point.}

\item{y}{Phenotype.}
}
\value{
Numeric evaluation.
}
\description{
Evaluate \eqn{K_{n}^{S}(t)}.
}
