% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-ExponentialCGF.R
\docType{data}
\name{ExponentialCGF}
\alias{ExponentialCGF}
\title{Exponential CGF Object}
\format{
An object of class \code{CGF} (R6), with usual methods \code{K},
\code{K1}, \code{K2}, \code{K3operator}, \code{K4operator}, etc.
}
\usage{
ExponentialCGF
}
\description{
A ready-to-use CGF object for the Exponential distribution.
This object is vectorized for i.i.d. replicates of the same rate.
}
\examples{
ExponentialCGF$K1(0, 2)  # (expected value for Exp(rate=2) is 1/2)

}
\keyword{datasets}
