% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-saddlepoint-mle.R
\name{find.zeroth.saddlepoint.MLE}
\alias{find.zeroth.saddlepoint.MLE}
\title{Find maximum likelihood estimates using the zeroth order saddlepoint approximation.}
\usage{
find.zeroth.saddlepoint.MLE(
  observed.data,
  model.cgf,
  starting.theta,
  lb.theta = rep(-Inf, times = length(starting.theta)),
  ub.theta = rep(Inf, times = length(starting.theta)),
  starting.tvec = rep(0, times = length(observed.data)),
  lb.tvec = rep(-Inf, times = length(observed.data)),
  ub.tvec = rep(Inf, times = length(observed.data)),
  std.error = FALSE,
  discrepancy = FALSE,
  user.ineq.constraint.function = NULL,
  opts.user = list(ftol_abs = 0, maxeval = 10000, xtol_rel = 1e-07, print_level = 0),
  sadd.eqn.opts = list(ftol_abs = 0, maxeval = 1000, xtol_rel = 1e-12, print_level = 0)
)
}
\arguments{
\item{observed.data}{a numeric vector - See TO DO list}

\item{model.cgf}{a CGF object corresponding to the observed data.}

\item{starting.theta}{a numeric vector of starting values for model parameters.}

\item{lb.theta}{a numeric vector, vector of lower bounds for model parameters, defaults to -Inf for all parameters.}

\item{ub.theta}{a numeric vector, vector of upper bounds for model parameters, defaults to Inf for all parameters.}

\item{starting.tvec}{a numeric vector, vector of starting values for saddlepoint parameters, defaults to zeros.}

\item{lb.tvec}{a numeric vector, vector of lower bounds for saddlepoint parameters, defaults to -Inf for all parameters.}

\item{ub.tvec}{a numeric vector, vector of upper bounds for saddlepoint parameters, defaults to Inf for all parameters.}

\item{std.error}{a logical value indicating whether to compute standard error for the estimates, defaults to FALSE.}

\item{discrepancy}{a logical value indicating whether to compute discrepancy on the estimates resulting from using saddlepoint likelihood, defaults to FALSE.}

\item{user.ineq.constraint.function}{a function, user-defined inequality constraint function, defaults to NULL.}

\item{opts.user}{A named list of options for the nloptr optimizer.}

\item{sadd.eqn.opts}{A named list of options for the nloptr optimizer. This sets options for optimization which are
only needed when \code{std.error = TRUE} or \code{discrepancy = TRUE}. By default, ftol_abs = 0, maxeval = 1e3, xtol_rel = 1.0e-12, print_level = 0.
Note: \code{xtol_rel} argument is set to 1.0e-12 and changing it may affect std.error computations.}
}
\value{
A list of MLEs. If \code{std.error} is TRUE, the list also includes standard errors of MLEs and inverse Hessian. See TO DO list
}
\description{
This function uses the nloptr package for optimization to find the maximum likelihood estimates (MLEs) using the zeroth order saddlepoint approximation. It takes the observed data, a CGF object that represents the CGF of the observed data, and various other parameters related to the optimization procedure.
}
\examples{
\dontrun{
# TO DO: Add examples
}
}
