% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint.R
\name{FindSaddlepoint}
\alias{FindSaddlepoint}
\title{Find Saddlepoint}
\usage{
FindSaddlepoint(g, mu, y, radius = 10, tol = 1e-06)
}
\arguments{
\item{g}{Genotype.}

\item{mu}{Mean of Y under the null.}

\item{y}{Phenotype.}

\item{radius}{Radius about the origin to search for saddlepoint.}

\item{tol}{Tolerance.}
}
\value{
Saddlepoint.
}
\description{
Calculate the saddlepoint: that \eqn{t} such that the first derivative of
the empirical cumulant generating function at \eqn{t} equals the observed
value of the score statistic.
}
