% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\name{fixed.parameter.adaptor}
\alias{fixed.parameter.adaptor}
\title{Create fixed parameter adaptor}
\usage{
fixed.parameter.adaptor(fixed_param)
}
\arguments{
\item{fixed_param}{Numeric vector of fixed parameters.}
}
\value{
An adaptor object set with fixed parameters.
}
\description{
Instantiates an adaptor with fixed parameter values. Useful for models where some parameters do not change.
}
\examples{
\dontrun{
# This is especially useful for setting known values of parameters that will not change during estimation.
# For instance, one might set the `n` parameter to `1` for a Bernoulli distribution (which is a special case of the binomial distribution).
}

}
\seealso{
\code{\link{indices.adaptor}}
}
