% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\name{fixed_parameter_adaptor}
\alias{fixed_parameter_adaptor}
\title{Create an adaptor using fixed values}
\usage{
fixed_parameter_adaptor(fixed_param)
}
\arguments{
\item{fixed_param}{A numeric vector indicating the fixed parameter values.}
}
\value{
"adaptor" object
}
\description{
Creates an adaptor using a specified numeric vector of fixed parameter values.
This is especially useful for setting known values of parameters that will not change during estimation.
For instance, one might set the \code{n} parameter to \code{1} for a Bernoulli distribution (which is a special case of the binomial distribution).
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{indices_adaptor}}
}
