% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionCGFs.R
\docType{data}
\name{GammaCGF}
\alias{GammaCGF}
\title{GammaCGF}
\format{
An object of class \code{"CGF"}.
}
\usage{
GammaCGF
}
\description{
A CGF object based on the Gamma distribution.
}
\examples{
\dontrun{
  K(tvec = 0, parameter.vector = c(10, 0.5), baseCGF = GammaCGF)
}

}
\keyword{datasets}
