% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-NormalCGF.R
\docType{data}
\name{NormalCGF}
\alias{NormalCGF}
\alias{GaussianCGF}
\title{Univariate Normal CGF object}
\format{
An object of class \code{VectorizedFunctionsCGF} (inherits from \code{CGF}, \code{R6}) of length 24.

An object of class \code{VectorizedFunctionsCGF} (inherits from \code{CGF}, \code{R6}) of length 24.
}
\usage{
NormalCGF

GaussianCGF
}
\description{
Ready-to-use CGF for a univariate Normal distribution with parameters
\eqn{(\mu,\sigma)}.
}
\details{
The parameter vector is interpreted as \code{c(mu, sigma)} (length 2),
or in vectorised form as \code{c(mu[1:L], sigma[1:L])}.

With \code{iidReps="any"} (the default), \code{length(tvec)} must be a multiple
of the number of parameter rows (L). In the scalar case (L=1), any length
\code{tvec} is treated as i.i.d. replicates.
}
\keyword{datasets}
