% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenoResid}
\alias{GenoResid}
\title{Genotypic Residual}
\usage{
GenoResid(g, x)
}
\arguments{
\item{g}{Genotype.}

\item{x}{Covariates.}
}
\value{
Numeric evaluation.
}
\description{
Residualize genotype with respect to covariates.
}
