% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{GeometricNonIdenticalModelCGF}
\alias{GeometricNonIdenticalModelCGF}
\title{Geometric Non-Identical Model-Based CGF}
\usage{
GeometricNonIdenticalModelCGF(prob)
}
\arguments{
\item{prob}{An object of class 'adaptor', representing the probability parameter in a Geometric distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Generates a CGF based on a non-identical Geometric distribution.
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
