% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\name{indices.adaptor}
\alias{indices.adaptor}
\title{Create index-based adaptor}
\usage{
indices.adaptor(indices)
}
\arguments{
\item{indices}{Numeric vector of positive integers for subsetting parameters.}
}
\value{
An adaptor object that subsets parameters by indices.
}
\description{
Creates an adaptor that subsets parameters from a vector using specified indices.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{fixed.parameter.adaptor}}
}
