% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\name{indices_adaptor}
\alias{indices_adaptor}
\title{Create an adaptor using indices}
\usage{
indices_adaptor(indices)
}
\arguments{
\item{indices}{A numeric vector of positive integers representing the subsetting indices}
}
\value{
"adaptor" object
}
\description{
Creates an adaptor using a specified numeric vector of positive integer indices.
These indices are for subsetting operations.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{fixed_parameter_adaptor}}
}
