% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ineqConstraint_with_gradient}
\alias{ineqConstraint_with_gradient}
\title{Compute inequality constraint and its gradient}
\usage{
ineqConstraint_with_gradient(combined_vector, ADfun_ineqConstraint)
}
\arguments{
\item{combined_vector}{A numeric vector of \code{tvec} and \code{theta} in that order.}

\item{ADfun_ineqConstraint}{An \code{XPtr} to a \verb{CppAD::ADFun<double>} object, typically the result of the \code{make.ADFunIneqConstraint()}-R function.}
}
\value{
A list with two elements: \code{fn}, the value of the inequality constraint function, and \code{gr}, the gradient of the function.
}
\description{
This function computes the inequality constraint function and its gradient for a given combined vector of \code{tvec} and \code{theta}.
}
\details{
This function is primarily intended for internal use.
}
\keyword{internal}
