% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-functions.R
\name{K}
\alias{K}
\title{K (CGF)}
\usage{
K(tvec, parameter.vector, baseCGF)
}
\arguments{
\item{tvec}{A numeric vector of t-values.}

\item{parameter.vector}{A numeric vector of parameters.}

\item{baseCGF}{A CGF type.}
}
\value{
A numeric of length 1.
}
\description{
Computes the Cumulant Genrating function, K for given t-vector, parameter vector, and a CGF object.
}
\examples{
\dontrun{
  K(tvec = c(0.003, 0.2), parameter.vector = c(10,0.5), baseCGF = BinomialCGF)
}
}
