% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-functions.R
\name{K1}
\alias{K1}
\title{K1}
\usage{
K1(tvec, parameter.vector, baseCGF)
}
\arguments{
\item{tvec}{A numeric vector of t-values.}

\item{parameter.vector}{A numeric vector of parameters.}

\item{baseCGF}{A CGF type.}
}
\value{
A numeric vector whose length equals the length of 'tvec'.
}
\description{
Computes the tvec-gradient of K for given t-values, a parameter vector, and a CGF object.
}
\examples{
\dontrun{
  K1(tvec = c(0.003, 0.2), parameter.vector = c(10,0.5), baseCGF = BinomialCGF)
}
}
