% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{K1_with_gradient}
\alias{K1_with_gradient}
\title{Compute K1 and its gradient}
\usage{
K1_with_gradient(combined_vector, ADfunK1)
}
\arguments{
\item{combined_vector}{A numeric vector of \code{tvec} and \code{theta} in that order.}

\item{ADfunK1}{An \code{XPtr} to a \verb{CppAD::ADFun<double>} object, typically the result of the \code{makeADFunK1()} function.}
}
\value{
A list with two elements: \code{fn}, the value of the K1 function, and \code{gr}, the gradient of the K1 function.
}
\description{
This function computes the K1 function and its gradient for a given combined vector of \code{tvec} and \code{theta}.
The K1 function is computed using a \verb{CppAD::ADFun<double>} object created by the \code{make.ADFunK1()}-R function.
}
\details{
This function is primarily intended for internal use.
}
\keyword{internal}
