% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-functions.R
\name{K2}
\alias{K2}
\title{K2}
\usage{
K2(tvec, parameter.vector, baseCGF)
}
\arguments{
\item{tvec}{A numeric vector of t-values.}

\item{parameter.vector}{A numeric vector of parameters.}

\item{baseCGF}{A CGF type.}
}
\value{
A square matrix with dimensions equivalent to the length of 'tvec'.
}
\description{
Computes the Hessian matrix of the function K with respect to 'tvec', for given t-values, a parameter vector, and a CGF object.
}
\examples{
\dontrun{
  K2(tvec = c(0.003, 0.2), parameter.vector = c(10,0.5), baseCGF = BinomialCGF)
}
}
