% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalCGF.R
\name{MultivariateNormalModelCGF}
\alias{MultivariateNormalModelCGF}
\title{Create a Multivariate Normal CGF Object}
\usage{
MultivariateNormalModelCGF(mu, sigma, iidReps = 1, ...)
}
\arguments{
\item{mu}{A function(\code{theta}) -> numeric vector (the mean).}

\item{sigma}{A function(\code{theta}) -> matrix (the covariance).}

\item{iidReps}{An integer. If >1, we replicate blocks for i.i.d. usage.}

\item{...}{Additional arguments passed to \code{\link{createCGF}} or
to \code{\link{iidReplicatesCGF}} if \code{iidReps>1}.}
}
\value{
A CGF object.
}
\description{
Creates a CGF for an arbitrary \code{d}-dimensional normal distribution.
The user supplies two functions:
\itemize{
\item \code{mu(theta)} returning a length-\eqn{d} numeric vector.
\item \code{sigma(theta)} returning a \eqn{d \times d} matrix.
}
}
\details{
\strong{Parameter Flattening}:
If \code{mu(theta)} is length \eqn{d}, and \code{Sigma(theta)} is \eqn{d x d},
then we store them as \code{c(mu, as.vector(Sigma))} in row-major or column-major form
(we just must be consistent in unwrapping).
}
