% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-functions.R
\name{neg_ll}
\alias{neg_ll}
\title{Compute the saddlepoint negative log-likelihood}
\usage{
neg_ll(tvec, parameter.vector, baseCGF)
}
\arguments{
\item{tvec}{A numeric vector of t-values at which the neg_ll is computed.}

\item{parameter.vector}{A numeric vector of parameters.}

\item{baseCGF}{A 'CGF' object which serves as the base for computing neg_ll.}
}
\value{
A numeric value, which is the computed saddlepoint negative log-likelihood.
}
\description{
The function computes the saddlepoint negative log-likelihood (neg_ll) at given t-values, a parameter vector, and a CGF object. See ... for MLEs
}
\examples{
\dontrun{
# Create a CGF object
# Compute neg_ll for given t-values and parameters
}
}
