% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negll_with_gradient}
\alias{negll_with_gradient}
\title{Compute negative log-likelihood and its gradient}
\usage{
negll_with_gradient(combined_vector, ADfun_negll)
}
\arguments{
\item{combined_vector}{A numeric vector of \code{tvec} and \code{theta} in that order.}

\item{ADfun_negll}{An \code{XPtr} to a \verb{CppAD::ADFun<double>} object, typically the result of the \code{make.ADFunNegll()}-R function.}
}
\value{
A list with two elements: \code{objective}, the value of the negative log-likelihood function, and \code{gradient}, the gradient of the function.
}
\description{
This function computes the negative log-likelihood function and its gradient for a given combined vector of \code{tvec} and \code{theta}.
}
\details{
This function is primarily intended for internal use.
}
\keyword{internal}
