% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-PoissonCGF.R
\name{PoissonModelCGF}
\alias{PoissonModelCGF}
\title{Create a Parametric Poisson CGF Object}
\usage{
PoissonModelCGF(lambda, iidReps = "any", ...)
}
\arguments{
\item{lambda}{A function or \code{adaptor} mapping \code{theta} -> scalar or vector of rates.}

\item{iidReps}{Either \code{"any"} or a positive integer. Default \code{"any"}.}

\item{...}{Passed through to the CGF creation (e.g., to override operators).}
}
\value{
A \code{CGF} object.
}
\description{
Poisson with rate(s) \code{lambda(theta)}. If \code{iidReps = "any"} (default),
\code{length(tvec)} must be a multiple of \code{length(lambda(theta))}. Supports i.i.d. and non-identical contexts
with optional length enforcement via \code{iidReps}. If \code{iidReps = m},
then \code{length(tvec)} must be \code{m * length(lambda(theta))}.
}
\examples{
# ex 1: i.i.d. scenario for univariate Poisson(rate = 2)
ex_iid <- PoissonModelCGF(lambda = function(x) x[1])
# OR ex_iid <- PoissonModelCGF(lambda = function(x) x[1], iidReps = 3)
ex_iid$K1(rep(0,3), 2)

# ex 2: non-identical scenario with replication: lambda returns c(2,5), iidReps=3
ex_repeat <- PoissonModelCGF(lambda = adaptor(indices = 1:2), iidReps = 3)
# OR ex_repeat <- PoissonModelCGF(lambda = function(x) c(2,5) )
ex_repeat$K1(rep(0,6), c(2,5))

}
