% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.saddlepoint}
\alias{print.saddlepoint}
\title{Print Method for Saddlepoint}
\usage{
\method{print}{saddlepoint}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{saddlepoint}.}

\item{digits}{Display digits.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{saddlepoint}.
}
