% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{saddlepoint-class}
\alias{saddlepoint-class}
\title{Saddlepoint}
\description{
Saddlepoint
}
\section{Slots}{

\describe{
\item{\code{ddksp}}{Second derivative of ECGF at the saddlepoint.}

\item{\code{dksp}}{First derivative of ECGF at the saddlepoint.}

\item{\code{ksp}}{ECGF at the saddlepoint.}

\item{\code{score}}{Score statistic.}

\item{\code{sp}}{Saddlepoint.}

\item{\code{w}}{Argument for saddlepoint approximation.}

\item{\code{v}}{Argument for saddlepoint approximation.}
}}

