% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,saddlepoint-method}
\alias{show,saddlepoint-method}
\title{Show Method for Saddlepoint Class}
\usage{
\S4method{show}{saddlepoint}(object)
}
\arguments{
\item{object}{An object of class \code{saddlepoint}.}
}
\description{
Show Method for Saddlepoint Class
}
