% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-package.R
\docType{package}
\name{saddlepoint-package}
\alias{saddlepoint}
\alias{saddlepoint-package}
\title{saddlepoint: A framework for the saddlepoint approximation with CGF operations}
\description{
Tools for saddlepoint approximations built around composable cumulant generating function (CGF) objects. Provides CGFs for common distributions, operators to combine CGFs (i.i.d. replication, sums, linear maps, random sums and tilts), and routines for saddlepoint likelihood inference including maximum likelihood estimation. Uses RTMB for automatic differentiation where derivatives are required.
}
\section{Package development}{

The package is still in development and there may still be bugs and errors.
While we do not expect major changes to the general user interface, there may be changes to the internal workings of the package as well as new additions and functionality.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://godrick.github.io/experiment-spa/}
  \item \url{https://github.com/godrick/experiment-spa}
}

}
\author{
\strong{Maintainer}: Godrick Oketch \email{godrick@berkeley.edu}

Authors:
\itemize{
  \item Jesse Goodman
  \item Rachel Fewster
}

}
\keyword{package}
