% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-functions.R
\name{saddlepoint.eqn.solve}
\alias{saddlepoint.eqn.solve}
\title{Solution of the saddlepoint equation}
\usage{
saddlepoint.eqn.solve(
  theta,
  y,
  cgf,
  starting.tvec = rep(0, times = length(y)),
  lb = rep(-Inf, times = length(y)),
  ub = rep(Inf, times = length(y)),
  sadd.eqn.opts = list(ftol_abs = 0, maxeval = 1000, xtol_rel = 1e-12, print_level = 0)
)
}
\arguments{
\item{theta}{A numeric vector of model parameters for which the CGF is defined.}

\item{y}{A numeric vector of observations.}

\item{cgf}{An object of type "CGF".}

\item{starting.tvec}{a numeric vector, vector of starting values for saddlepoint parameters, defaults to zeros.}

\item{lb}{a numeric vector, vector of lower bounds for saddlepoints tvec, defaults to -Inf.}

\item{ub}{a numeric vector, vector of upper bounds for saddlepoints tvec, defaults to Inf.}

\item{sadd.eqn.opts}{A list of options for the nlopt optimizer. This should be a named list, where the names are option names and the list elements are the corresponding option values.
The default options are: \code{list(algorithm="NLOPT_LD_SLSQP", ftol_abs = 0, maxeval = 1e3, xtol_rel = 1.0e-12, print_level = 0)}.
Note that the option "algorithm" cannot be changed. See the documentation for the \code{nloptr} package for more details about the options.}
}
\value{
The optimal value of \code{t.vec} for which K1(t.vec, theta) = y, according to the nlopt optimization.
}
\description{
This function minimises the expression \eqn{K'(t, \theta)-y} of the saddlepoint equation, and returns
the value of \eqn{t} for which \eqn{K'(t, \theta) = y}, for a specified/known value of \eqn{\theta}.
}
\examples{
# TO DO: Add examples

}
