% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saddlepoint-package.R
\docType{package}
\name{Saddlepoint}
\alias{Saddlepoint}
\title{Saddlepoint: Saddlepoint Approximation}
\description{
Association score test with p-value calculation via the empirical saddle
point approximation.
}
\author{
Zachary R. McCaw
}
