% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{scalarToiidCGF}
\alias{scalarToiidCGF}
\title{Transforms a scalar CGF to an i.i.d CGF}
\usage{
scalarToiidCGF(scalarCGF)
}
\arguments{
\item{scalarCGF}{A CGF object representing the scalar CGF to be transformed.}
}
\value{
A CGF object representing the resulting i.i.d CGF.
}
\description{
Transforms a scalar CGF to an i.i.d CGF
}
