% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-test.R
\name{ScoreTest}
\alias{ScoreTest}
\title{Score Test}
\usage{
ScoreTest(g, mu, sigma2, y, alpha = 0.05, radius = 10, tol = 1e-06, x = NULL)
}
\arguments{
\item{g}{Genotypic residuals.}

\item{mu}{Mean of Y under the null.}

\item{sigma2}{Residual variance under the null.}

\item{y}{Phenotype.}

\item{alpha}{Type I error.}

\item{radius}{Radius about the origin to search for saddlepoint.}

\item{tol}{Tolerance.}

\item{x}{Null model covariates, if not already regressed from genotype.}
}
\value{
Data.frame containing the score and p-value.
}
\description{
Perform a score test of the null hypothesis that genotype is unrelated
to the phenotype, where the p-value is calculated by empirical
saddlepoint approximation.
}
