% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalCGF-creatingFunctions.R
\name{sourceCppInSaddlepointEnv}
\alias{sourceCppInSaddlepointEnv}
\title{Compile and load externally created 'CGF' objects.}
\usage{
sourceCppInSaddlepointEnv(file, cacheDir = getwd(), ...)
}
\arguments{
\item{file}{A character string giving the directory and name of the file to be compiled and loaded.}

\item{cacheDir}{A character string giving the directory to be used for caching compiled code.
If not provided, the current working directory will be used.}

\item{...}{Other arguments passed on to \code{Rcpp::sourceCpp()}.}
}
\value{
A list containing all functions exported using \verb{// [[Rcpp::export]]}. Each object is associated with its name in the list.\code{CGF_with_AD} objects will be of class 'CGF'.
}
\description{
This function compiles and loads 'CGF' objects and associated functions that are not predefined in the 'saddlepoint' package.
The C++ code should be written following the provided template.
}
