% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalCGF.R
\name{UnivariateNormalModelCGF}
\alias{UnivariateNormalModelCGF}
\title{Create a Parametric Gaussian CGF Object}
\usage{
UnivariateNormalModelCGF(mean, sd, iidReps = 1, ...)
}
\arguments{
\item{mean}{A function that accepts a single parameter vector \code{theta} and returns \eqn{\mu}.}

\item{sd}{A function that accepts a single parameter vector \code{theta} and returns \eqn{\sigma}.}

\item{iidReps}{A positive integer. If \code{iidReps>1},
we expect \code{tvec} to have that length.}

\item{...}{Additional arguments passed to base CGF creation function.}
}
\value{
A CGF object.
}
\description{
Constructs a CGF object for a univariate normal distribution where mean \eqn{\mu(\theta)}
and variance \eqn{\sigma(\theta)} are derived from user-supplied logic.
The user supplies a function that maps their full parameter vector
\code{theta} to \eqn{(\mu, \sigma)}.
}
