% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{adaptCGF}
\alias{adaptCGF}
\title{adaptCGF}
\usage{
adaptCGF(baseCGF, adaptor)
}
\arguments{
\item{baseCGF}{A 'CGF' object to be adapted}

\item{adaptor}{An 'adaptor' object created by one of the adaptor creating functions.}
}
\value{
A CGF object.
}
\description{
Adapts a CGF object using a specified adaptor.
}
\details{
The function create...
}
\examples{
\dontrun{
  # Adapting the BinomialCGF with a fixed parameter adaptor
  adaptedCGF <- adapt.CGF(baseCGF = BinomialCGF,
                          adaptor = adaptorUsingFixedParam(c(10, 0.5)) )

  # Creating a Binomial Model CGF with the same parameters
  modelCGF <- BinomialModelCGF(n = adaptorUsingFixedParam(10),
                               prob = adaptorUsingFixedParam(0.5))

  # Testing if the adaptedCGF and modelCGF give the same result when used as baseCGF in function K
  res1 <- K(tvec = 0.003, parameter.vector = c(10,0.5), baseCGF = adaptedCGF)
  res2 <- K(tvec = 0.003, parameter.vector = c(10,0.5), baseCGF = modelCGF)
  print(all.equal(res1, res2)) # Should print TRUE
}
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
