% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-AdaptCGF.R
\name{adaptCGF}
\alias{adaptCGF}
\title{Create an adapted CGF object using an adaptor function}
\usage{
adaptCGF(cgf, param_adaptor, ...)
}
\arguments{
\item{cgf}{A \code{CGF} object}

\item{param_adaptor}{A function with signature \code{function(theta) -> adapted_param}.}

\item{...}{Additional arguments passed to \code{createCGF()} (rarely needed).}
}
\value{
A new CGF object (class \code{"CGF"}).
}
\description{
Takes an existing CGF and a parameter adaptor function, then returns
a new CGF object whose methods first adapt the parameter vector before
calling the original CGF's methods.
}
\examples{
# Suppose we have a PoissonCGF object with parameter = lambda.
# We want to treat the first element of a bigger vector as lambda:
## adaptor <- function(param) param[1]
## new_cgf <- adaptCGF(PoissonCGF, adaptor)
## new_cgf$K(0.1, c(2, 99))  # Internally calls PoissonCGF$K(0.1, 2).

}
