% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-class.R
\name{adaptor}
\alias{adaptor}
\title{adaptor Constructor}
\usage{
adaptor(ptr)
}
\arguments{
\item{ptr}{An externalptr object. It should not be an empty externalptr object.}
}
\value{
An object of class "adaptor".
}
\description{
This function constructs an object of class "adaptor" using a valid externalptr object.
}
\examples{
\dontrun{
  # assuming ptr is a valid externalptr object
  adaptor <- adaptor(ptr)
}
}
\seealso{
\code{\link{CGF}}
}
