% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionCGFs.R
\docType{data}
\name{BinomialCGF}
\alias{BinomialCGF}
\title{BinomialCGF}
\format{
An object of class \code{"CGF"}.
}
\usage{
BinomialCGF
}
\description{
A CGF object based on the Binomial distribution.
}
\examples{
\dontrun{
  # Using the BinomialCGF in the K function
  K(tvec = 0, parameter.vector = c(10, 0.5), baseCGF = BinomialCGF)
}

}
\keyword{datasets}
