% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{BinomialCGF}
\alias{BinomialCGF}
\title{BinomialCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
BinomialCGF
}
\description{
A CGF object for the binomial distribution that includes the CGF and its related derivatives.
When using this object, specify the \code{parameter_vector} as a vector with
the number of trials \code{n} and the probability of success \code{p} in that order.
}
\examples{
\dontrun{
  # Compute the CGF and its first derivative at t = 0 for n = 10 trials and p = 0.5 
  cgf_value = BinomialCGF$K(tvec = 0, parameter_vector = c(10, 0.5))
  first_derivative = BinomialCGF$K1(tvec = 0, parameter_vector = c(10, 0.5))
  }
}
\keyword{datasets}
