% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-BinomialCGF.R
\docType{data}
\name{BinomialCGF}
\alias{BinomialCGF}
\title{Binomial CGF Object}
\format{
An object of class \code{CGF} (R6), with standard methods:
\code{K}, \code{K1}, \code{K2}, \code{K3operator}, etc.
}
\usage{
BinomialCGF
}
\description{
A ready-to-use CGF object for the Binomial distribution. This object expects
calls of the form \code{BinomialCGF$K(tvec, param)}, where \code{param}
is a numeric vector \eqn{(n, p)}:
\describe{
\item{n}{A positive integer (the number of trials).}
\item{p}{A probability in \eqn{(0,1)}.}
}
}
\details{
By default, \code{BinomialCGF} supports vectorized evaluation for i.i.d. replicates of
Binomial random variables.
}
\examples{
# Evaluate the expected value of X ~ Binomial(10, 0.3) using the BinomialCGF object
BinomialCGF$K1(0, c(10, 0.3))

}
\keyword{datasets}
