% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{BinomialModelCGF}
\alias{BinomialModelCGF}
\title{BinomialModelCGF}
\usage{
BinomialModelCGF(n, prob)
}
\arguments{
\item{n}{An object of class 'adaptor', representing the parameter n in a Binomial distribution.}

\item{prob}{An object of class 'adaptor', representing the probability parameter in a Binomial distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Constructs a CGF for the Binomial distribution with flexible parameter specifications using adaptors. The adaptors offer multiple methods to specify parameters:
\enumerate{
\item Using indices from an external parameter vector (\code{\link{adaptorUsingIndices}}).
\item Setting fixed values that won't change during model computations (\code{\link{adaptorUsingFixedParam}}).
\item Employing R functions for dynamic computation of parameters (\code{\link{adaptorUsingRFunctions}}).
}

This adaptability facilitates model constructions where specific parameters might be varied during computations while keeping others constant.
}
\examples{
\dontrun{
  # Using adaptorUsingFixedParam for n and adaptorUsingIndices for prob
  cgf = BinomialModelCGF(n = adaptorUsingFixedParam(10), prob = adaptorUsingIndices(1))
  # Compute CGF with: K(tvec = ..., parameter.vector = 0.5, baseCGF = cgf)
}

}
