% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-compute-saddlepointLL-correction.R
\name{compute.saddlepointLL.correction}
\alias{compute.saddlepointLL.correction}
\title{Compute the correction term for the saddlepoint approximation to the log-likelihood}
\usage{
compute.saddlepointLL.correction(
  parameter_vector,
  observed.data,
  cgf,
  tvec.hat = NULL,
  gradient = FALSE,
  hessian = FALSE,
  spa_method = "standard",
  ...
)
}
\arguments{
\item{parameter_vector}{Numeric vector of parameters.}

\item{observed.data}{Numeric vector of observed data.}

\item{cgf}{A \code{CGF} object.}

\item{tvec.hat}{(Optional) Numeric vector. If supplied, \code{tvec} is taken directly
as this vector (the saddlepoint \eqn{\hat{t}}). Otherwise, we solve for it numerically if \code{cgf$has_analytic_tvec_hat()} is \code{FALSE}.}

\item{gradient}{Logical. If \code{TRUE}, the gradient wrt \code{parameter_vector} is returned.}

\item{hessian}{Logical. If \code{TRUE}, the Hessian wrt \code{parameter_vector} is returned.}

\item{spa_method}{Character string. One of \code{"standard"} or \code{"zeroth"}.}

\item{...}{Optional arguments to be passed to \code{\link{saddlepoint.solve}}.}
}
\value{
A named list with elements:
\describe{
\item{val}{Numeric scalar of the correction term.}
\item{gradient}{The gradient at \code{theta}, if \code{gradient=TRUE}.}
\item{hessian}{The Hessian at \code{theta}, if \code{hessian=TRUE}.}
}
}
\description{
This function calculates the correction term for the saddlepoint approximation
to the log-likelihood. By default, it computes that of the \strong{first-order} approximation,
If \code{spa_method="zeroth"}, it instead computes the correction term for the \strong{zeroth-order} approximation.
}
\seealso{
\code{\link{find.saddlepoint.MLE}}, \code{\link{compute.spa.negll}}
}
