% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-compute-spa-negll.R
\name{compute.spa.negll}
\alias{compute.spa.negll}
\title{Compute a saddlepoint negative log-likelihood}
\usage{
compute.spa.negll(
  parameter_vector,
  observed.data,
  cgf,
  tvec.hat = NULL,
  gradient = FALSE,
  hessian = FALSE,
  method = "standard",
  ...
)
}
\arguments{
\item{parameter_vector}{Numeric vector of parameters.}

\item{observed.data}{Numeric vector of observed data.}

\item{cgf}{A \code{CGF} object.}

\item{tvec.hat}{(Optional) numeric vector. If provided, it overrides other ways
of finding \code{tvec}.}

\item{gradient}{Logical. If \code{TRUE}, the gradient is computed.}

\item{hessian}{Logical. If \code{TRUE}, the Hessian is computed.}

\item{method}{Character string. One of \code{"standard"} or \code{"zeroth"}.}

\item{...}{Additional arguments passed to \code{saddlepoint.solve()} if \code{tvec.hat} is not provided.}
}
\value{
A named list with elements:
\describe{
\item{vals}{Numeric scalar. The negative log-likelihood at \code{parameter_vector}.}
\item{gradient}{Numeric vector. The gradient at \code{parameter_vector}, if \code{gradient=TRUE}.}
\item{hessian}{The Hessian at \code{parameter_vector}, if \code{hessian=TRUE}.}
}
}
\description{
Computes either "standard" or "zeroth-order" saddlepoint approximation to the negative log-likelihood.
}
\seealso{
\code{\link{saddlepoint.solve}}
}
