% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-functions.R
\name{compute.std.error}
\alias{compute.std.error}
\title{Compute standard error and inverse Hessian}
\usage{
compute.std.error(
  observed.data,
  combined.estimates,
  model.cgf,
  objective.function,
  lb.tvec = rep(-Inf, times = length(observed.data)),
  ub.tvec = rep(Inf, times = length(observed.data)),
  sadd.eqn.opts = list(ftol_abs = 0, maxeval = 1000, xtol_rel = 1e-12, print_level = 0)
)
}
\arguments{
\item{observed.data}{A numeric vector of observed data}

\item{combined.estimates}{TO DO A numeric vector of combined estimated parameters}

\item{model.cgf}{A CGF object}

\item{objective.function}{A function that defines the objective to be minimized...TO DO}

\item{lb.tvec}{A numeric vector specifying the lower bounds for tvec (default: -Inf)}

\item{ub.tvec}{A numeric vector specifying the upper bounds for tvec (default: Inf)}

\item{sadd.eqn.opts}{A list of options for the nlopt optimizer. This should be a named list, where the names are option names and the list elements are the corresponding option values.
The default options are: \code{list(algorithm="NLOPT_LD_SLSQP", ftol_abs = 0, maxeval = 1e3, xtol_rel = 1.0e-12, print_level = 0)}.
Note that the option "algorithm" cannot be changed. See the documentation for the \code{nloptr} package for more details about the options.}
}
\value{
A list containing the standard error (std.error) and the inverse Hessian matrix (inverse.hessian)
}
\description{
This function calculates the standard error and the inverse of the Hessian matrix ....
}
\examples{
# TODO: Add examples

}
