% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-functions.R
\name{compute.std.error}
\alias{compute.std.error}
\title{Compute standard error and inverse Hessian}
\usage{
compute.std.error(
  observed.data,
  estimated.tvec,
  estimated.theta,
  cgf,
  zeroth.order = FALSE,
  non.saddlepoint.negll.function = NULL
)
}
\arguments{
\item{observed.data}{A numeric vector of observations used in the saddlepoint likelihood calculations.}

\item{estimated.tvec}{A numeric vector of MLE values for the saddlepoint values \code{tvec}.}

\item{estimated.theta}{A numeric vector of MLE values for the model parameters.}

\item{cgf}{A CGF object used in the saddlepoint likelihood computation.}

\item{zeroth.order}{A logical value indicating whether the zeroth-order saddlepoint likelihood is used. Default is FALSE.}

\item{non.saddlepoint.negll.function}{An optional function specifying a non-saddlepoint negative log-likelihood. If your model used for estimation incorporates a likelihood component that is not based on the saddlepoint approximation, this function must be provided. See details for more information.}
}
\value{
A list containing:
\itemize{
\item std.error: Standard error computed from the diagonal of the inverse Hessian.
\item inverse.hessian: Inverse of the Hessian matrix.
}
}
\description{
This function calculates the standard error and the inverse of the Hessian matrix
for a model incorporating both the saddlepoint likelihood and an optional
non-saddlepoint negative log-likelihood component.
}
\details{
The \code{non.saddlepoint.negll.function}, if provided, should have the form:
\deqn{function(theta) \{ return(list(objective = ..., hessian = ...)) \}}
It adds flexibility by allowing incorporation of likelihood components not based on the saddlepoint likelihood.
It is important when the analysis combines the saddlepoint likelihood with an external likelihood.
If not provided, the function defaults to using solely the saddlepoint likelihood.
}
\examples{
# TODO: Add examples

}
