% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-ConcatenationCGF.R
\name{concatenationCGF}
\alias{concatenationCGF}
\title{Concatenation of CGF objects.}
\usage{
concatenationCGF(
  cgf_list,
  component_dims = 1,
  block_size = NULL,
  iidReps = NULL,
  ...
)
}
\arguments{
\item{cgf_list}{A non-empty list of \code{CGF} objects.}

\item{component_dims}{An integer vector (same length as \code{cgf_list}) specifying the length of \code{tvec} compatible with each \code{CGF}. Defaults to 1 for each \code{CGF}.}

\item{block_size}{Optional integer specifying the block size for i.i.d. replication of the resulting \code{CGF}. Defaults to \code{NULL}.}

\item{iidReps}{Optional integer specifying the number of i.i.d. blocks that will be expected the resulting \code{CGF}. Defaults to \code{NULL}.}

\item{...}{Additional arguments passed to the \code{CGF} creation function.}
}
\value{
A \code{CGF} object.
}
\description{
Constructs a CGF object for a concatenated vector of independent random variables,
with each represented by its respective CGF object and a corresponding dimension.
This means CGF \code{cgf_list[[i]]} sees \code{component_dims[i]} consecutive components of \code{tvec}.
It facilitates the modeling of a concatenated random vector,
where each segment of the vector is specified by its CGF object and a corresponding dimension.
}
\details{
\itemize{
\item The dimension of \code{tvec} is \code{sum(component_dims)}.
\item If \code{iidReps} or \code{block_size} is specified, the result is automatically wrapped
via \code{iidReplicatesCGF()}.
}
}
\examples{
\dontrun{
# 
}
}
\seealso{
\code{\link{iidReplicatesCGF}}, \code{\link{sumOfIndependentCGF}}
}
