% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{concatenationCGF}
\alias{concatenationCGF}
\title{Concatenation CGF}
\usage{
concatenationCGF(cgfWithLengths)
}
\arguments{
\item{cgfWithLengths}{A list where each entry consists of a CGF object and its associated vector length. Each element is a sublist containing:
\itemize{
\item The CGF object (CGF object).
\item The associated vector length for that CGF (positive integer).
}}
}
\value{
A CGF object.
}
\description{
Enables the combination of multiple CGFs into a concatenated structure where each CGF has an associated vector length.
}
\examples{
\dontrun{
# TO DO
}

}
