% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-class.R
\name{createCGF}
\alias{createCGF}
\title{Create a CGF object}
\usage{
createCGF(ptr)
}
\arguments{
\item{ptr}{An external pointer 'externalptr' expected by the CGF class.}
}
\value{
An object of class 'CGF'.
}
\description{
This function initializes a new CGF object with a valid external pointer.
}
