% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-CGF_base.R
\name{createCGF}
\alias{createCGF}
\title{A factory function to create a CGF object from user-defined functions}
\usage{
createCGF(
  K,
  K1,
  K2,
  K3operator,
  K4operator,
  ineq_constraint = NULL,
  param_adaptor = function(x) x,
  analytic_tvec_hat_func = NULL,
  op_name = "UnnamedOperation",
  tilting_exponent = NULL,
  neg_ll = NULL,
  func_T = NULL,
  K4operatorAABB = NULL,
  K3K3operatorAABBCC = NULL,
  K3K3operatorABCABC = NULL,
  K4operatorAABB_factored = NULL,
  K3K3operatorAABBCC_factored = NULL,
  K3K3operatorABCABC_factored = NULL,
  K2operator = NULL,
  K2operatorAK2AT = NULL,
  ...
)
}
\arguments{
\item{K}{A function implementing K(tvec, parameter_vector) returning a scalar.}

\item{K1}{A function implementing K1(tvec, parameter_vector) returning a numeric vector.}

\item{K2}{A function implementing K2(tvec, parameter_vector) returning a numeric matrix.}

\item{K3operator}{A function implementing ...}

\item{K4operator}{A function implementing ...}

\item{ineq_constraint}{(optional) If supplied, overrides the default \code{ineq_constraint}.}

\item{param_adaptor}{(optional) A function: param_adaptor(parameter_vector) returning adapted parameters.
Default is the identity function.}

\item{analytic_tvec_hat_func}{(Optional) function to compute tvec using analytical formula.}

\item{op_name}{A character string indicating the operation or transformation being performed.
Default is "UnnamedOperation".}

\item{tilting_exponent}{(optional) If supplied, overrides the default \code{tilting_exponent}.}

\item{neg_ll}{(optional) If supplied, overrides the default \code{neg_ll}.}

\item{func_T}{(optional) If supplied, overrides the default \code{func_T}.}

\item{K4operatorAABB}{(optional) If supplied, overrides the default \code{K4operatorAABB}.}

\item{K3K3operatorAABBCC}{(optional) If supplied, overrides the default \code{K3K3operator}.}

\item{K3K3operatorABCABC}{(optional) If supplied, overrides the default \code{K3K3operatorABCABC}.}

\item{K4operatorAABB_factored}{(optional) If supplied, overrides the default \code{K4operatorAABB_factored}.}

\item{K3K3operatorAABBCC_factored}{(optional) If supplied, overrides the default \code{K3K3operatorAABBCC_factored}.}

\item{K3K3operatorABCABC_factored}{(optional) If supplied, overrides the default \code{K3K3operatorABCABC_factored}.}

\item{K2operator}{(optional) If supplied, overrides the default \code{K2operator}.}

\item{K2operatorAK2AT}{(optional) If supplied, overrides the default \code{K2operatorAK2AT}.}

\item{...}{Additional named optional methods that can be passed to the CGF object.}
}
\value{
An object of class \code{CGF}.
}
\description{
\code{createCGF()} creates an object that represents a CGF using user-supplied functions,
instead of requiring the user to define a subclass. This approach is more flexible
and avoids the need for inheritance in many cases.
}
