% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-CGF_fromVectorizedFunctions.R
\name{createCGF_fromVectorisedFunctions}
\alias{createCGF_fromVectorisedFunctions}
\title{Create a \code{CGF} object from vectorized functions}
\usage{
createCGF_fromVectorisedFunctions(
  K_vectorized_func,
  K1_vectorized_func,
  K2_vectorized_func,
  K3_vectorized_func,
  K4_vectorized_func,
  ineq_constraint_func = NULL,
  analytic_tvec_hat_func = NULL,
  op_name = "UnnamedOperation",
  tilting_exponent = NULL,
  neg_ll = NULL,
  func_T = NULL,
  rsim = NULL,
  K4operatorAABB = NULL,
  K3K3operatorAABBCC = NULL,
  K3K3operatorABCABC = NULL,
  K4operatorAABB_factored = NULL,
  K3K3operatorAABBCC_factored = NULL,
  K3K3operatorABCABC_factored = NULL,
  K2operator = NULL,
  K2operatorAK2AT = NULL,
  ...
)
}
\arguments{
\item{K_vectorized_func}{A function of the form \code{function(tvec, param) -> numeric vector} that returns the CGF values.}

\item{K1_vectorized_func}{A function of the form \code{function(tvec, param) -> numeric vector} that returns the first derivative values.}

\item{K2_vectorized_func}{A function of the form \code{function(tvec, param) -> numeric vector} that returns the second derivative values.}

\item{K3_vectorized_func, K4_vectorized_func}{Similar vectorized functions for the third and fourth derivatives.}

\item{ineq_constraint_func}{Optional inequality constraint function.}

\item{analytic_tvec_hat_func}{Optional tvec_hat function override.}

\item{op_name}{Optional character string indicating the name of the operation or transformation being performed.}

\item{tilting_exponent}{Optional tilting exponent function override.}

\item{neg_ll}{Optional neg_ll function override.}

\item{func_T}{Optional func_T function override.}

\item{rsim}{Optional simulation method. A function of the form
\code{function(n, vector_length, parameter_vector, tvec = NULL, ...)} returning
a numeric vector or matrix of length \code{n * vector_length} or a
\code{vector_length x n} matrix.}

\item{K4operatorAABB, K3K3operatorAABBCC, K3K3operatorABCABC}{Optional operator overrides.}

\item{K4operatorAABB_factored, K3K3operatorAABBCC_factored, K3K3operatorABCABC_factored}{Optional factored operator overrides.}

\item{K2operator, K2operatorAK2AT}{Optional operator overrides.}

\item{...}{Any additional named optional methods.}
}
\value{
A \code{CGF} object.
}
\description{
This function allows you to create a \code{CGF} object using vectorized functions
along with any optional operators or methods.
}
