% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF_fromVectorizedFunctions.R
\name{createCGF_fromVectorisedFunctions}
\alias{createCGF_fromVectorisedFunctions}
\title{Create a \code{CGF} object from vectorized functions}
\usage{
createCGF_fromVectorisedFunctions(
  K_vectorized_func,
  K1_vectorized_func,
  K2_vectorized_func,
  K3_vectorized_func,
  K4_vectorized_func,
  ineq_constraint_func = NULL,
  param_adaptor = function(x) x,
  analytic_tvec_hat_func = NULL,
  op_name = "UnnamedOperation",
  tilting_exponent = NULL,
  neg_ll = NULL,
  func_T = NULL,
  K4operatorAABB = NULL,
  K3K3operatorAABBCC = NULL,
  K3K3operatorABCABC = NULL,
  K4operatorAABB_factored = NULL,
  K3K3operatorAABBCC_factored = NULL,
  K3K3operatorABCABC_factored = NULL,
  K2operator = NULL,
  K2operatorAK2AT = NULL,
  ...
)
}
\arguments{
\item{K_vectorized_func}{Function: returns a vector of K-values for tvec.}

\item{K1_vectorized_func}{Function: returns a vector (gradient) for K1.}

\item{K2_vectorized_func}{Function: returns a vector of second derivatives (often diagonal for IID).}

\item{K3_vectorized_func, K4_vectorized_func}{Similar vectorized functions for K3 and K4.}

\item{ineq_constraint_func}{Optional inequality constraint function.}

\item{param_adaptor}{Optional parameter adaptor function (default identity).}

\item{analytic_tvec_hat_func}{Optional tvec_hat function override.}

\item{op_name}{Optional character string indicating the name of the operation or transformation being performed.}

\item{tilting_exponent}{Optional tilting exponent function override.}

\item{neg_ll}{Optional neg_ll function override.}

\item{func_T}{Optional func_T function override.}

\item{K4operatorAABB, K3K3operatorAABBCC, K3K3operatorABCABC}{Optional operator overrides.}

\item{K4operatorAABB_factored, K3K3operatorAABBCC_factored, K3K3operatorABCABC_factored}{Optional factored operator overrides.}

\item{K2operator, K2operatorAK2AT}{Optional operator overrides.}

\item{...}{Any additional optional methods.}
}
\value{
A \code{CGF} object.
}
\description{
This function allows you to create a \code{CGF} object using vectorized functions
along with any optional operators or methods.
}
