% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionCGFs.R
\docType{data}
\name{ExponentialCGF}
\alias{ExponentialCGF}
\title{ExponentialCGF}
\format{
An object of class \code{"CGF"}.
}
\usage{
ExponentialCGF
}
\description{
A CGF object based on the Exponential distribution.
}
\examples{
\dontrun{
  # Using the ExponentialCGF in the K function
  K(tvec = 0, parameter.vector = 10, baseCGF = ExponentialCGF)
}

}
\keyword{datasets}
