% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-ExponentialCGF.R
\docType{data}
\name{ExponentialCGF}
\alias{ExponentialCGF}
\title{Exponential CGF Object}
\format{
An object of class \code{CGF} (R6), with typical methods like
\code{K}, \code{K1}, \code{K2}, \code{K3operator}, \code{K4operator}, etc.
}
\usage{
ExponentialCGF
}
\description{
A ready-to-use CGF object for the Exponential distribution.
}
\details{
\strong{Parameter Vector}:
The parameter vector must be a single positive numeric value representing the rate parameter.

\strong{Cumulant Generating Function}:
\deqn{
  K(t) = -\log(\lambda - t) + \log(\lambda), \quad t < \lambda.
}

By default, \code{ExponentialCGF} is vectorized for i.i.d. replicates. If you supply a
\code{tvec} of length \emph{m}, it sums the CGF across \emph{m} i.i.d. exponentials,
each with parameter \eqn{\lambda}.
}
\examples{
# Evaluate K, K1, K2 at t=0 for rate=2
#  -> K(0)= -log(2-0)+log(2)=  -log(2)+log(2)=0
ExponentialCGF$K(0, 2)

#  -> K1(0)= 1/(2-0)= 0.5  (the mean for Exp with rate=2 is 1/2)
ExponentialCGF$K1(0, 2)

#  -> K2(0)= 1/(2-0)^2= 0.25
ExponentialCGF$K2(0, 2)

}
\keyword{datasets}
