% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{ExponentialCGF}
\alias{ExponentialCGF}
\title{ExponentialCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
ExponentialCGF
}
\description{
An CGF object containing the CGF and its related derivatives for the exponential distribution
}
\examples{
\dontrun{
  mean_value = ExponentialCGF$K1(tvec = 0, parameter_vector = 15)
  }
}
\keyword{datasets}
