% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{ExponentialModelCGF}
\alias{ExponentialModelCGF}
\title{Exponential Model-Based CGF}
\usage{
ExponentialModelCGF(rate)
}
\arguments{
\item{rate}{An object of class 'adaptor', representing the rate parameter in the Exponential distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Generates a CGF based on the Exponential distribution.
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
