% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-ExponentialCGF.R
\name{ExponentialModelCGF}
\alias{ExponentialModelCGF}
\title{Create a Parametric Exponential CGF Object}
\usage{
ExponentialModelCGF(rate, iidReps = 1, ...)
}
\arguments{
\item{rate}{A function(\code{theta}) -> numeric, returning a positive rate parameter \eqn{\lambda}.
\itemize{
\item Optionally, you can pass an \strong{adaptor} object.
\item In many cases, an \strong{adaptor} object can be used instead (especially if \eqn{\lambda} is constant or index-based).
}}

\item{iidReps}{A positive integer specifying the number of i.i.d. replicates
the CGF handles in \code{tvec}. Defaults to 1. If \code{iidReps>1}, \code{tvec} must be length \eqn{iidReps}.}

\item{...}{Additional arguments passed to the base CGF creation function, such as optional method overrides.}
}
\value{
An object of class \code{CGF}
}
\description{
Constructs a CGF object for the Exponential distribution where the rate parameter
\eqn{\lambda(\theta)} is derived from a user-supplied function.
}
\examples{
rate_func <- function(theta) theta[1]  # e.g. param -> 2
expo_cgf <- ExponentialModelCGF(rate=rate_func, iidReps=1)
# Evaluate K1 at t=0 with param=c(2)
expo_cgf$K1(0, c(2)) 

}
