% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-saddlepoint-mle.R
\name{find.saddlepoint.MLE}
\alias{find.saddlepoint.MLE}
\title{Find maximum likelihood estimates using the saddlepoint likelihood.}
\usage{
find.saddlepoint.MLE(
  observed.data,
  cgf,
  starting.theta,
  lb.theta = rep(-Inf, times = length(starting.theta)),
  ub.theta = rep(Inf, times = length(starting.theta)),
  starting.tvec = rep(0, times = length(observed.data)),
  lb.tvec = rep(-Inf, times = length(observed.data)),
  ub.tvec = rep(Inf, times = length(observed.data)),
  std.error = FALSE,
  discrepancy = FALSE,
  user.ineq.constraint.function = NULL,
  opts.user = list(ftol_abs = 0, maxeval = 10000, xtol_rel = 1e-07, print_level = 0),
  zeroth.order = FALSE
)
}
\arguments{
\item{observed.data}{A numeric vector, matrix, or data frame containing the observed data.}

\item{cgf}{A CGF object corresponding to the distribution of the observed data.}

\item{starting.theta}{A numeric vector of starting values for model parameters.}

\item{lb.theta}{A numeric vector of lower bounds for \code{theta}. Defaults to \code{-Inf}.}

\item{ub.theta}{A numeric vector of upper bounds for \code{theta}. Defaults to \code{Inf}.}

\item{starting.tvec}{A numeric vector of starting values for the saddlepoint \eqn{t}. Defaults to \code{0}.}

\item{lb.tvec}{A numeric vector, vector of lower bounds for saddlepoint parameters, defaults to -Inf for all parameters.}

\item{ub.tvec}{A numeric vector, vector of upper bounds for saddlepoint parameters, defaults to Inf for all parameters.}

\item{std.error}{Logical. If \code{TRUE}, standard errors of the MLEs are computed. Defaults to \code{FALSE}.}

\item{discrepancy}{Logical. If \code{TRUE}, compute a discrepancy measure that approximates
how different the saddlepoint approximation-based estimates are from the unknown true MLEs. Defaults to \code{FALSE}.}

\item{user.ineq.constraint.function}{Optional user-defined inequality constraints. Defaults to \code{NULL}.}

\item{opts.user}{A named list of options for the nloptr optimizer. This
should be a subset of the following elements: ftol_abs (> 0),
maxeval (a positive integer), xtol_rel (> 0), print_level (0, 1, 2, or 3).
See the nloptr package documentation for more details about these options.
By default, ftol_abs = 0, maxeval = 1e4, xtol_rel = 1.0e-7, print_level = 0.
Note: The algorithm option is fixed to "NLOPT_LD_SLSQP" and cannot be changed by the user.}

\item{zeroth.order}{A logical value indicating whether to use the zeroth-order saddlepoint likelihood. Defaults to FALSE, using the standard saddlepoint likelihood.}
}
\value{
A list containing:
\itemize{
\item \code{MLEs.tvec, MLEs.theta}: The estimated saddlepoint \eqn{\hat{t}} and parameter vector \eqn{\hat{\theta}}.
\item \code{std.error}, \code{inverse.hessian}: If \code{std.error=TRUE}, these hold the standard errors and the inverse Hessian.
\item \code{discrepancy}: If \code{discrepancy=TRUE}, this is the approximated discrepancy between the resulting saddlepoint approximation-based MLEs and the unknown true MLEs.
\item \code{solution, status, message}: The raw output from \code{nloptr}.
}
}
\description{
This function uses the nloptr package for optimization to find the maximum likelihood estimates (MLEs).
It takes a CGF object along with various parameters related to the optimization procedure.
An alternative zeroth-order method can be enabled by setting the \code{zeroth.order} parameter to \code{TRUE}.
}
\details{
\itemize{
\item \strong{Observed Data}: \code{observed.data} can be a numeric vector or a matrix/data frame.
If it's a matrix/data frame, columns are treated as i.i.d. replicate blocks.
\item \strong{Zeroth-order}: The \code{zeroth.order} argument allows toggling between the standard and zeroth-order saddlepoint likelihoods.
When \code{zeroth.order} is set to TRUE, the objective function is modified to the zeroth-order saddlepoint approximation to the negative log-likelihood
\deqn{\hat{t} y - K_Y (\hat{t}; \theta).}
The estimates of \eqn{\theta} are calculated based on observations \eqn{y},
where \eqn{K_Y(t; \theta)} is the CGF of the underlying random variable, and \eqn{\hat{t}} is the solution
to the saddlepoint equation. When FALSE, the standard saddlepoint likelihood is used.
}
}
\examples{
\dontrun{
# TO DO: Add examples
}
}
