% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-saddlepoint-mle.R
\name{find.saddlepoint.MLE}
\alias{find.saddlepoint.MLE}
\title{Find maximum likelihood estimates using the saddlepoint likelihood.}
\usage{
find.saddlepoint.MLE(
  observed.data,
  cgf,
  starting.theta,
  lb.theta = rep(-Inf, times = length(starting.theta)),
  ub.theta = rep(Inf, times = length(starting.theta)),
  starting.tvec = rep(0, times = length(observed.data)),
  lb.tvec = rep(-Inf, times = length(observed.data)),
  ub.tvec = rep(Inf, times = length(observed.data)),
  std.error = FALSE,
  discrepancy = FALSE,
  user.ineq.constraint.function = NULL,
  opts.user = list(ftol_abs = 0, maxeval = 10000, xtol_rel = 1e-07, print_level = 0),
  zeroth.order = FALSE
)
}
\arguments{
\item{observed.data}{a numeric vector - See TO DO list}

\item{cgf}{a CGF object corresponding to the observed data.}

\item{starting.theta}{a numeric vector of starting values for model parameters.}

\item{lb.theta}{a numeric vector, vector of lower bounds for model parameters, defaults to -Inf for all parameters.}

\item{ub.theta}{a numeric vector, vector of upper bounds for model parameters, defaults to Inf for all parameters.}

\item{starting.tvec}{a numeric vector, vector of starting values for saddlepoint parameters, defaults to zeros.}

\item{lb.tvec}{a numeric vector, vector of lower bounds for saddlepoint parameters, defaults to -Inf for all parameters.}

\item{ub.tvec}{a numeric vector, vector of upper bounds for saddlepoint parameters, defaults to Inf for all parameters.}

\item{std.error}{a logical value indicating whether to compute standard error for the estimates, defaults to FALSE.}

\item{discrepancy}{a logical value indicating whether to compute discrepancy on the estimates resulting from using saddlepoint likelihood, defaults to FALSE.}

\item{user.ineq.constraint.function}{a function, user-defined inequality constraint function, defaults to NULL.}

\item{opts.user}{A named list of options for the nloptr optimizer. This
should be a subset of the following elements: ftol_abs (> 0),
maxeval (a positive integer), xtol_rel (> 0), print_level (0, 1, 2, or 3).
See the nloptr package documentation for more details about these options.
By default, ftol_abs = 0, maxeval = 1e4, xtol_rel = 1.0e-7, print_level = 0.
Note: The algorithm option is fixed to "NLOPT_LD_SLSQP" and cannot be changed by the user.}

\item{zeroth.order}{A logical value indicating whether to use the zeroth-order saddlepoint likelihood. Defaults to FALSE, using the standard saddlepoint likelihood.}
}
\value{
A list of MLEs. If \code{std.error} is TRUE, the list also includes standard errors of MLEs and inverse Hessian. See TO DO list
}
\description{
This function uses the nloptr package for optimization to find the maximum likelihood estimates (MLEs). It takes a CGF object that represents the CGF for a random variable, corresponding to the observed data, along with various parameters related to the optimization procedure. An alternative zeroth-order method can be enabled by setting the \code{zeroth.order} parameter to TRUE.
}
\details{
The \code{zeroth.order} argument allows toggling between the standard and zeroth-order saddlepoint likelihoods.
When \code{zeroth.order} is set to TRUE, the objective function is modified to the zeroth-order saddlepoint negative log-likelihood
\deqn{\hat{t} y - K_Y (\hat{t}; \theta).}
Here, the estimates of \eqn{\theta} are calculated based on observations \eqn{y},
where \eqn{K_Y(t; \theta)} is the CGF of the underlying random variable, and \eqn{\hat{t}} is the solution
to the saddlepoint equation. When FALSE, the standard saddlepoint likelihood is used.
}
\examples{
\dontrun{
# TO DO: Add examples
}
}
