% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{GammaCGF}
\alias{GammaCGF}
\title{GammaCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
GammaCGF
}
\description{
An CGF object containing the CGF and its related derivatives for the gamma distribution
}
\examples{
\dontrun{
  ## For X~Geometric(shape = 7.5, rate = 2), E(X) is  
  GammaCGF$K1(tvec = 0, parameter_vector = c(7.5, 2)) 
  }
}
\keyword{datasets}
