% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaCGF.R
\docType{data}
\name{GammaCGF}
\alias{GammaCGF}
\title{Gamma CGF Object}
\format{
An object of class \code{CGF} (an R6 class), with the usual methods:
\code{K, K1, K2, K3operator, K4operator}, etc.
}
\usage{
GammaCGF
}
\description{
A ready-to-use CGF object for the Gamma distribution with shape \eqn{\alpha}
and rate \eqn{\beta}. The \code{parameter_vector} used when calling methods like \code{K(tvec, parameter_vector)}
shoulde be a numeric vector \eqn{c(\alpha, \beta)}.
}
\details{
\strong{CGF}: For a Gamma random variable \eqn{X} with shape \eqn{\alpha} and rate
\eqn{\beta}, the cumulant generating function is:
\deqn{K(t;\alpha, \beta) = -\alpha \,\log \bigl(1 - t/\beta\bigr), \quad t < \beta.}

\strong{Parameter Vector}: The \code{parameter_vector} is assumed to have the form
\eqn{(\alpha, \beta)}. You must ensure
that \eqn{t < \beta} for all evaluations. This object enforces that constraint
as an inequality \eqn{t - \beta < 0}.
}
\examples{
# Evaluate K at t=0.5 for shape=2, rate=2 (thus t<2).
# param = c(2, 2)
GammaCGF$K(0.5, c(2,2))

}
\keyword{datasets}
