% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{GammaModelCGF}
\alias{GammaModelCGF}
\title{GammaCGF with flexible parameterization}
\usage{
GammaModelCGF(shape, rate)
}
\arguments{
\item{shape}{Adaptor or function representing the shape parameter of the gamma distribution.}

\item{rate}{Adaptor or function representing the rate parameter of the gamma distribution.}
}
\value{
'CGF' object configured for the gamma distribution.
}
\description{
Constructs a CGF object for the gamma distribution, supporting flexible parameter specifications through adaptors.
This function allows for dynamic adjustments of the (\code{shape}) and (\code{rate}) parameters using similar mechanisms as described for the binomial distribution.
For detailed explanations on using adaptors for dynamic parameter adjustments, see \code{\link{BinomialModelCGF}}.
}
\examples{
\dontrun{
## Example using a fixed parameter for `rate` and an index adaptor for `shape`
# `rate` parameter fixed at 1
# `shape` will be part of a parameter vector that will be accessed using the the first index
cgf <- GammaModelCGF(shape = indices.adaptor(1), rate = fixed.parameter.adaptor(1))
cgf$K1(tvec = 0, parameter_vector = 0.2) == GammaCGF$K1(tvec = 0, parameter_vector = c(0.2, 1))
}
}
\seealso{
\code{\link{GammaCGF}} \code{\link{BinomialModelCGF}}
}
