% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{GammaModelCGF}
\alias{GammaModelCGF}
\title{GammaModelCGF}
\usage{
GammaModelCGF(shape, rate)
}
\arguments{
\item{shape}{An object of class 'adaptor', representing the parameter shape in a Gamma distribution.}

\item{rate}{An object of class 'adaptor', representing the rate parameter in a Gamma distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Constructs a CGF for the Gamma distribution with parameter specifications using adaptors. The adaptors offer multiple methods to specify parameters:
\enumerate{
\item Using indices from an external parameter vector (\code{\link{adaptorUsingIndices}}).
\item Setting fixed values that won't change during model computations (\code{\link{adaptorUsingFixedParam}}).
\item Employing R functions for dynamic computation of parameters (\code{\link{adaptorUsingRFunctions}}).
}
}
\examples{
\dontrun{

}

}
