% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-GeometricCGF.R
\docType{data}
\name{GeometricCGF}
\alias{GeometricCGF}
\title{Geometric CGF Object}
\format{
An object of class \code{CGF} (R6), with usual methods:
\code{K, K1, K2, K3operator, K4operator}, etc.
}
\usage{
GeometricCGF
}
\description{
A ready-to-use CGF object for a single-parameter Geometric distribution.
This object correspond to the random variable that counts the number of failures before achieving the first success.
The \code{parameter_vector} for \code{GeometricCGF} is taken to be the success probability \eqn{prob}.
By default, this object is vectorized for i.i.d. replicates of Geometric random variables.
}
\examples{
# the expected value of X~Geometric(prob = 0.3) using the CGF 
GeometricCGF$K1(0, 0.3)

}
\seealso{
\code{\link{GeometricModelCGF}}
}
\keyword{datasets}
