% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{GeometricCGF}
\alias{GeometricCGF}
\title{GeometricCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
GeometricCGF
}
\description{
An CGF object containing the CGF and its related derivatives for the geometric distribution
}
\examples{
\dontrun{
  ## For X~Geometric(0.03), E(X) is  
  GeometricCGF$K1(tvec = 0, parameter_vector = 0.03) 
  }
}
\keyword{datasets}
