% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionCGFs.R
\docType{data}
\name{GeometricCGF}
\alias{GeometricCGF}
\title{GeometricCGF}
\format{
An object of class \code{"CGF"}.
}
\usage{
GeometricCGF
}
\description{
A CGF object based on the Geometric distribution.
}
\examples{
\dontrun{
  # Using the GeometricCGF in the K function
  K(tvec = 0, parameter.vector = 0.03, baseCGF = GeometricCGF) # Note that tvec must be less than -log(1-parameter.vector)
}

}
\keyword{datasets}
