% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{GeometricModelCGF}
\alias{GeometricModelCGF}
\title{Geometric Model-Based CGF}
\usage{
GeometricModelCGF(prob)
}
\arguments{
\item{prob}{An object of class 'adaptor', representing the probability parameter in a Geometric distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Generates a CGF based on the Geometric distribution.
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
