% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-functions.R
\name{get.zeroth.saddlepoint.nll.function}
\alias{get.zeroth.saddlepoint.nll.function}
\title{Create the zeroth-order saddlepoint negative log-likelihood function}
\usage{
get.zeroth.saddlepoint.nll.function(tvec, theta, cgf)
}
\arguments{
\item{tvec}{A numeric vector.}

\item{theta}{A numeric vector.}

\item{cgf}{An object of class 'CGF'.}
}
\value{
A function that takes a vector 'a' as an argument. When \code{a = c(tvec, theta)} is passed to the returned function, it yields a list in the format: \code{list(objective = , gradient = )}.
}
\description{
This function creates and returns a function of the form \code{function(a) \{...\}}, where 'a' combines \code{tvec} and \code{theta} arguments to a single vector, in that order.
}
\examples{
\dontrun{
  TO DO: write a working example
}

}
