% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-IIDReplicatesCGF.R
\name{iidReplicatesCGF}
\alias{iidReplicatesCGF}
\title{Replicate a CGF over multiple i.i.d blocks or with a fixed block size}
\usage{
iidReplicatesCGF(cgf, iidReps = NULL, block_size = NULL, ...)
}
\arguments{
\item{cgf}{A CGF object.}

\item{iidReps}{Either NULL or a positive integer specifying the number of IID blocks.}

\item{block_size}{Either NULL or a positive integer specifying the size of each block.}

\item{...}{Additional arguments passed to \code{\link{createCGF}} in case
you want to override any methods.}
}
\value{
A new CGF object that operates on length-\eqn{m d} input vectors, chunked
into \eqn{m} blocks, each block of length \eqn{d}.
}
\description{
Extends a given CGF to handle multiple IID blocks based on either \code{iidReps}
or a fixed \code{block_size}. Exactly one of these parameters must be provided.
}
\details{
\itemize{
\item If \code{iidReps} is provided, the input vector is split into \code{iidReps} equal-sized blocks.
\item If \code{block_size} is provided, the number of blocks is determined by the length of the input vector.
\item Exactly one of \code{iidReps} or \code{block_size} should be non-NULL.
}
}
\examples{
# Suppose cgf is dimension=3, and we want 5 blocks => dimension=15
#   aggregator <- iidReplicatesCGF(cgf, iidReps=5)

}
