% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{iidReplicatesCGF}
\alias{iidReplicatesCGF}
\title{Transforms a CGF object to an i.i.d-based CGF object.}
\usage{
iidReplicatesCGF(cgf, block_size)
}
\arguments{
\item{cgf}{(CGF object): The CGF for the random vector of dimension (block_size).}

\item{block_size}{(positive integer): The vector length of a single replicate.}
}
\value{
A CGF object.
}
\description{
Transforms a CGF object to an i.i.d-based CGF object.
}
