% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{linearlyMappedCGF}
\alias{linearlyMappedCGF}
\title{CGF of a linearly mapped random variable.}
\usage{
linearlyMappedCGF(cgf, matrix_A)
}
\arguments{
\item{cgf}{The CGF object for the random variable before the linear transformation.}

\item{matrix_A}{The transformation matrix applied to the random variable.}
}
\value{
A CGF object
}
\description{
Computes the CGF object of a linearly transformed random variable.
}
