% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{linearlyMappedCGF}
\alias{linearlyMappedCGF}
\title{CGF of a linearly mapped random variable.}
\usage{
linearlyMappedCGF(baseCGF, matrix_A)
}
\arguments{
\item{baseCGF}{The CGF object representing the random variable before the linear transformation.}

\item{matrix_A}{The transformation matrix applied to the random variable.}
}
\value{
A CGF object representing the transformed random variable.
}
\description{
Evaluates the CGF object of a linearly transformed random variable.
}
