% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomialCGF.R
\docType{data}
\name{MultinomialCGF}
\alias{MultinomialCGF}
\title{MultinomialCGF}
\format{
An object of class \code{CGF} (an R6 class), with standard methods
\code{K}, \code{K1}, \code{K2}, \code{K3operator}, \code{K4operator}, etc.
}
\usage{
MultinomialCGF
}
\description{
A ready-to-use CGF object for the multinomial distribution.
}
\details{
\strong{Parameter Vector Interpretation}:
The \code{parameter_vector} used when calling methods like \code{K(tvec, parameter_vector)}
should be of the form \eqn{(N, x_1, \dots, x_d)}, where \eqn{N} is the total count,
and \eqn{x = x_i \dots x_d} can be any vector of non-zero numbers not all zero.
Therefore, \eqn{x} can be the multinomial probability vector, and can also be interpreted as a vector of odds.
}
\examples{
tvec <- c(0.1, -0.2, 0.05)
param <- c(10, 2, 3, 5) # total count=10, odds=(2,3,5)
MultinomialCGF$K(tvec, param)

}
\keyword{datasets}
