% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-MultinomialCGF.R
\docType{data}
\name{MultinomialCGF}
\alias{MultinomialCGF}
\title{MultinomialCGF}
\format{
An object of class \code{CGF} (an R6 class), with standard methods
\code{K}, \code{K1}, \code{K2}, \code{K3operator}, \code{K4operator}, etc.
}
\usage{
MultinomialCGF
}
\description{
A ready-to-use CGF object for a single multinomial random vector. For a more general object that handles i.i.d. cases, use \code{\link{MultinomialModelCGF}}.
}
\details{
\strong{Parameter Vector Interpretation}:
The \code{parameter_vector} used when calling methods like \code{K1(tvec, parameter_vector)}
should be of the form \eqn{(N, x_1, \dots, x_d)}, where \eqn{N} is the total count,
and \eqn{x = x_i \dots x_d} can be any vector of non-negative entries, not all zero.
Therefore, \eqn{x} can be the multinomial probability vector, and can also be interpreted as a vector of odds.
}
\examples{
tvec <- rep(0,3)
parameter_vector <- c(10, 2, 3, 5)         # total count=10, odds=(2,3,5) or probabilities=c(0.2,0.3,0.5)
MultinomialCGF$K1(tvec, parameter_vector)  # MultinomialCGF$K1(tvec, c(0.2,0.3,0.5))

}
\keyword{datasets}
