% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionCGFs.R
\docType{data}
\name{MultinomialCGF}
\alias{MultinomialCGF}
\title{MultinomialCGF}
\format{
An object of class \code{"CGF"}.
}
\usage{
MultinomialCGF
}
\description{
A CGF object based on the Multinomial distribution.
}
\examples{
\dontrun{
  # Using the MultinomialCGF in the K function
  K(tvec = rep(0,3), parameter.vector = c(10, 0.1, 0.2, 0.7), baseCGF = MultinomialCGF)
}

}
\keyword{datasets}
