% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{MultinomialCGF}
\alias{MultinomialCGF}
\title{MultinomialCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
MultinomialCGF
}
\description{
An CGF object containing the CGF and its related derivatives for the multinomial distribution
}
\examples{
\dontrun{
  ## For X~Multinomial(n = 10, prob = c(0.1, 0.2, 0.3, 0.4)), E(X) is  
  MultinomialCGF$K1(tvec = rep(0,4), parameter_vector = c(10, 0.1, 0.2, 0.3, 0.4)) 
  }
}
\keyword{datasets}
