% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{MultinomialModelCGF}
\alias{MultinomialModelCGF}
\title{Multinomial Model-Based CGF}
\usage{
MultinomialModelCGF(n, prob_vec)
}
\arguments{
\item{n}{An object of class 'adaptor'.}

\item{prob_vec}{An object of class 'adaptor', representing a vector of Multinomial probabilities.}
}
\value{
An object of class 'CGF'.
}
\description{
Generates a CGF based on the Multinomial distribution.
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
