% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-MultinomialCGF.R
\name{MultinomialModelCGF}
\alias{MultinomialModelCGF}
\title{Create a Parametric MultinomialCGF Object}
\usage{
MultinomialModelCGF(N, prob_vec, iidReps = 1, ...)
}
\arguments{
\item{N}{A function of the form \code{function(theta) -> numeric}, returning the total count \eqn{N}.}

\item{prob_vec}{A function of the form \code{function(theta) -> numeric vector}, returning
the probabilities/odds for the multinomial distribution.}

\item{iidReps}{Integer. If \code{iidReps > 1}, we replicate the resulting CGF for
multiple i.i.d. copies of the multinomial variables. Defaults to \code{1}.}

\item{...}{Additional arguments passed to \code{\link{createCGF}} or possibly
to \code{\link{iidReplicatesCGF}} if \code{iidReps > 1}.}
}
\value{
A CGF object.
}
\description{
Constructs a CGF object for the multinomial distribution with a user-specified
parameter mapping. Optionally, you can replicate this CGF for \code{iidReps}
independent and identically distributed (i.i.d.) observations.
}
