% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-MultinomialCGF.R
\name{MultinomialModelCGF}
\alias{MultinomialModelCGF}
\title{Create a Parametric MultinomialCGF Object}
\usage{
MultinomialModelCGF(n, prob_vec, block_size = NULL, iidReps = NULL, ...)
}
\arguments{
\item{n}{A function of the form \code{function(theta) -> numeric}, returning the total count \eqn{n}.}

\item{prob_vec}{A function of the form \code{function(theta) -> numeric vector}, returning
the probabilities/odds for the multinomial distribution.}

\item{block_size}{Either \code{NULL} or a positive integer specifying the size of each block
for i.i.d. replication. Defaults to \code{NULL}.}

\item{iidReps}{Either \code{NULL} or a positive integer specifying how many i.i.d. blocks
to expect. Defaults to \code{NULL}.}

\item{...}{Additional named arguments passed to \code{\link{createCGF}}.}
}
\value{
A CGF object.
}
\description{
Constructs a CGF object for the multinomial distribution with a user-specified
parameter mapping. Optionally, you can replicate this CGF for \code{iidReps}
independent and identically distributed (i.i.d.) observations.
}
