% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{NegBinCGF}
\alias{NegBinCGF}
\title{NegBinCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 9.
}
\usage{
NegBinCGF
}
\description{
A CGF object for the negative binomial distribution.
When using this object, specify the \code{parameter_vector} as a vector with the number of successes \code{r} and the probability of success \code{p} in that order.
}
\keyword{datasets}
