% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-NegBinCGF.R
\docType{data}
\name{NegBinCGF}
\alias{NegBinCGF}
\title{Negative Binomial CGF Object}
\format{
An object of class \code{CGF} (an R6 class), with the usual methods:
\code{K, K1, K2, K3operator, K4operator}, etc.
}
\usage{
NegBinCGF
}
\description{
A ready-to-use CGF object for the Negative Binomial distribution with number of successes \eqn{r}
and success probability parameter \eqn{p}. The \code{parameter_vector} used when calling methods such as \code{K(tvec, parameter_vector)}
should be a numeric vector \eqn{c(r, p)}.
}
\keyword{datasets}
