% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-NegBinCGF.R
\name{NegBinModelCGF}
\alias{NegBinModelCGF}
\title{Create a Parametric Negative Binomial CGF Object}
\usage{
NegBinModelCGF(r, p, iidReps = "any", ...)
}
\arguments{
\item{r}{A function (or \code{adaptor})  that accepts a single parameter vector \code{theta} and returns the number of successes.}

\item{p}{A function (or \code{adaptor})  that accepts a single parameter vector \code{theta} and returns a scalar success probability or a vector of success probabilities.}

\item{iidReps}{Either \code{"any"} or a positive integer specifying how many
i.i.d. blocks are expected. Defaults to \code{"any"}, meaning no restriction on the length of \code{tvec}.}

\item{...}{Additional arguments passed to the underlying CGF creation function.}
}
\value{
A \code{CGF} object.
}
\description{
Creates a CGF object for the Negative Binomial distribution with number of successes \eqn{r(\theta)} and
success probability parameter \eqn{p(\theta)} defined by user-provided parameter functions.
This function supports both i.i.d. and non-identical usage.
}
