% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{NegBinModelCGF}
\alias{NegBinModelCGF}
\title{Create a Parametric NegBin CGF object}
\usage{
NegBinModelCGF(r, p)
}
\arguments{
\item{r}{Adaptor or function representing the number of successes parameter of the Negative Binomial distribution.}

\item{p}{Adaptor or function representing the success probability parameter of the Negative Binomial distribution.}
}
\value{
'CGF' object.
}
\description{
Creates a CGF for the Negative Binomial distribution with the parameters \code{r} and \code{p} specified as adaptors or by user-supplied function.
}
\seealso{
\code{\link{NegBinCGF}}
}
