% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalCGF.R
\docType{data}
\name{NormalCGF}
\alias{NormalCGF}
\title{Univariate Normal CGF object}
\format{
An object of class \code{CGF} (R6), with the usual methods \code{K, K1, K2, etc.}
}
\usage{
NormalCGF
}
\description{
A ready-to-use CGF object for the univariate Normal distribution with mean
\eqn{\mu} and standard deviation \eqn{\sigma}.
}
\details{
\strong{Parameter Vector}: The \code{parameter_vector} should be
\eqn{(\mu, \sigma)} in that order.

If \code{tvec} has length \eqn{n}, we interpret that as
\eqn{n} i.i.d. replicates.
}
\examples{
# Example: Evaluate K at tvec=c(0.1, -0.2) for mu=2, var=4
# param = c(2, 4)
#### TODO: Add example

}
\keyword{datasets}
