% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-PoissonCGF.R
\docType{data}
\name{PoissonCGF}
\alias{PoissonCGF}
\title{Poisson CGF Object}
\format{
An object of class \code{VectorizedFunctionsCGF} (inherits from \code{CGF}, \code{R6}) of length 20.
}
\usage{
PoissonCGF
}
\description{
Ready-to-use CGF for Poisson. Accepts scalar or vector \code{lambda}.
If \code{length(tvec)} is a multiple of \code{length(lambda)}, evaluation proceeds (iidReps="any").
}
\keyword{datasets}
