% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCGFs.R
\docType{data}
\name{PoissonCGF}
\alias{PoissonCGF}
\title{PoissonCGF}
\format{
An object of class \code{CGF} (inherits from \code{R6}) of length 8.
}
\usage{
PoissonCGF
}
\description{
An CGF object containing the CGF and its related derivatives for a binomial distribution
}
\examples{
\dontrun{
  # Access the CGF for a Poisson distribution with lambda = 15
  cgf_value = PoissonCGF$K(tvec = 0.1, parameter_vector = 15)
  }
}
\keyword{datasets}
