% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptedCGFs.R
\name{PoissonModelCGF}
\alias{PoissonModelCGF}
\title{Poisson Model-Based CGF}
\usage{
PoissonModelCGF(lambda)
}
\arguments{
\item{lambda}{An object of class 'adaptor', representing the rate parameter lambda in a Poisson distribution.}
}
\value{
An object of class 'CGF'.
}
\description{
Generates a CGF based on the Poisson distribution.
}
\seealso{
\code{\link{adaptorUsingRFunctions}}, \code{\link{adaptorUsingIndices}}, \code{\link{adaptorUsingFixedParam}}
}
