% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{randomlyStoppedSumCGF}
\alias{randomlyStoppedSumCGF}
\title{CGF of a randomly stopped sum.}
\usage{
randomlyStoppedSumCGF(count_cgf, summand_cgf)
}
\arguments{
\item{count_cgf}{The CGF object representing the random variable \eqn{N}.}

\item{summand_cgf}{The CGF object representing the random variables \eqn{\tilde{X}_i}.}
}
\value{
A CGF object of the random variable \eqn{\tilde{Y}}.
}
\description{
Computes the CGF object for a randomly stopped sum \eqn{\tilde{Y} = \sum_{i=1}^{N} \tilde{X}_i}{Y = sum(X[1], X[2], ..., X[N])}, where \eqn{N} and \eqn{\tilde{X}_i} are random variables. Here, \eqn{\tilde{X}_i} are i.i.d. random variables independent of \eqn{N}, which is a non-negative integer-valued random variable.
}
\details{
The setup is as follows:
\deqn{{\tilde{Y} = \sum_{i=1}^{N} \tilde{X}_i}}
The random variable \eqn{\tilde{Y}} is computed as the sum of \eqn{N} i.i.d copies of the random vector \eqn{\tilde{X}_i}. Each \eqn{\tilde{X}_i} is a random vector of arbitrary dimension \eqn{d}.
For example, when \eqn{d=3}:
\deqn{ \tilde{X_1} = (X_{11}, X_{12}, X_{13}) }
\deqn{ \tilde{X_2} = (X_{21}, X_{22}, X_{23}) }
\deqn{ ... }
\deqn{ \tilde{X_N} = (X_{N1}, X_{N2}, X_{N3}) }
Here, \eqn{X_{i1}, X_{i2}, X_{i3}} are scalars but the vectors \eqn{\tilde{X_i}} are i.i.d copies.
}
