% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-package.R
\docType{package}
\name{saddlepoint-package}
\alias{saddlepoint}
\alias{saddlepoint-package}
\title{saddlepoint: A Framework for the Saddlepoint Approximation with CGF Operations}
\description{
Tools for saddlepoint approximations built around composable cumulant generating function (CGF) objects. Provides CGFs for common distributions, operators to combine CGFs (i.i.d. replication, sums, linear maps, random sums and tilts), and routines for saddlepoint likelihood inference including maximum likelihood estimation. Uses RTMB for automatic differentiation where derivatives are required. The framework is described in Oketch, Fewster and Goodman (2026) \doi{10.48550/arXiv.2607.17464}.
}
\section{Package development}{

The package is still in development and there may still be bugs and errors.
While we do not expect major changes to the general user interface, there may be changes to the internal workings of the package as well as new additions and functionality.
}

\references{
Oketch, G., Fewster, R. M., and Goodman, J. (2026).
\emph{A general framework for computation and estimation using the saddlepoint
approximation}. \doi{10.48550/arXiv.2607.17464}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://godrick.github.io/saddlepoint/}
  \item \url{https://github.com/godrick/saddlepoint}
}

}
\author{
\strong{Maintainer}: Godrick Oketch \email{godrick@berkeley.edu}

Authors:
\itemize{
  \item Rachel M. Fewster
  \item Jesse Goodman
}

}
\keyword{package}
