% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-package.R
\docType{package}
\name{saddlepoint-package}
\alias{saddlepoint}
\alias{saddlepoint-package}
\title{saddlepoint: Saddlepoint methods with CGF operations}
\description{
Implements saddlepoint approximations based on cumulant generating functions (CGFs). Distributions are represented as CGF objects that provide derivatives needed for saddlepoint approximation methods and can be combined using compositional operators (e.g., i.i.d. replication, linear maps, and sums). Includes routines for solving saddlepoint equations and for maximum likelihood estimation, with optional automatic differentiation support via RTMB.
}
\section{Package development}{

The package is still in development and there may still be bugs and errors.
While we do not expect major changes to the general user interface, there may be changes to the internal workings of the package as well as new additions and functionality.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://godrick.github.io/experiment-spa/}
  \item \url{https://github.com/godrick/experiment-spa}
}

}
\author{
\strong{Maintainer}: Godrick Oketch \email{godrick@berkeley.edu}

Authors:
\itemize{
  \item Jesse Goodman
  \item Rachel Fewster
}

}
\keyword{package}
