% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-package.R
\docType{package}
\name{saddlepoint-package}
\alias{saddlepoint}
\alias{saddlepoint-package}
\title{saddlepoint: What the Package Does (One Line, Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs in the Description.
}
\keyword{internal}
