% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint-package.R
\docType{package}
\name{saddlepoint-package}
\alias{saddlepoint}
\alias{saddlepoint-package}
\title{saddlepoint: Saddlepoint methods with CGF operations}
\description{
This package provides a comprehensive framework for performing operations on Cumulant Generating Functions (CGF) and efficiently computing saddlepoint approximation-related functions for statistical modeling. It enables users to manage various operations on random variables through their CGFs and to compute saddlepoint-based maximum likelihood estimates.
}
\section{Package development}{

The package is still in development and there may still be bugs and errors.
While we do not expect major changes to the general user interface, there may be changes to the internal workings of the package as well as new additions and functionality.
}

\author{
\strong{Maintainer}: Godrick Oketch \email{goke316@aucklanduni.ac.nz}

Authors:
\itemize{
  \item Jesse Goodman
  \item Rachel Fewster
}

}
\keyword{package}
