% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-SubunitaryMultinomialCGF.R
\name{SubunitaryMultinomialModelCGF}
\alias{SubunitaryMultinomialModelCGF}
\title{SubunitaryMultinomialModelCGF}
\usage{
SubunitaryMultinomialModelCGF(
  n,
  prob_vec,
  block_size = NULL,
  iidReps = NULL,
  ...
)
}
\arguments{
\item{n}{An \code{adaptor} or a function of the form \code{function(theta) -> numeric}, returning the number of trials in the multinomial distribution.}

\item{prob_vec}{An \code{adaptor} or a function of the form \code{function(theta) -> numeric vector}, returning
the multinomial probabilities. The sum of these probabilities should be strictly less than one to validate the occurrence of the known zero outcome.}

\item{block_size}{Either \code{NULL} or a positive integer specifying the size of each block
for i.i.d. replication. Defaults to \code{NULL}.}

\item{iidReps}{Either \code{NULL} or a positive integer specifying how many i.i.d. blocks
to expect. Defaults to \code{NULL}.}

\item{...}{Additional named arguments passed to \code{\link{createCGF}}.}
}
\value{
A \eqn{CGF} object.
}
\description{
Generates a CGF object derived from a modified Multinomial distribution. This modification accommodates instances where certain outcomes are explicitly zero, and consequently, the probabilities associated with the remaining non-zero outcomes sum to a value less than one.
}
\details{
Given \eqn{x_1,...,x_k} as the non-zero outcomes of a Multinomial distribution that sum to \eqn{N},
with each outcome having an associated probability \eqn{p_i} where \eqn{ \sum_{i=1}^{k} p_i < 1},
this CGF is designed to integrate the occurrence of a known zero outcome, \eqn{x_{k+1}}, thereby ensuring \eqn{ \sum_{i=1}^{k+1} p_i = 1}.

This approach allows for the modeling and analysis of Multinomial distributions in situations where some categories are known not to be observed and the probabilities of observing the other categories are adjusted accordingly.
}
\seealso{
\code{\link{MultinomialModelCGF}}
}
