% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{sumOfiidCGF}
\alias{sumOfiidCGF}
\title{CGF of a sum of 'n' i.i.d random variables.}
\usage{
sumOfiidCGF(baseCGF, n)
}
\arguments{
\item{baseCGF}{A CGF object the single random variable.}

\item{n}{An integer representing the number of i.i.d random variables to be summed.}
}
\value{
A CGF object representing the sum of 'n' i.i.d random variables.
}
\description{
Computes the CGF object for the sum of 'n' i.i.d random variables.
}
