% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{sumOfiidCGF}
\alias{sumOfiidCGF}
\title{Sum of 'n' i.i.d random variables.}
\usage{
sumOfiidCGF(cgf, n)
}
\arguments{
\item{cgf}{A CGF object for the random variable to be summed.}

\item{n}{An integer representing the number of i.i.d random variables to be summed.}
}
\value{
A CGF object.
}
\description{
Computes the CGF object for the sum of 'n' i.i.d random variables.
}
