% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-SumOfIndependentCGF.R
\name{sumOfIndependentCGF}
\alias{sumOfIndependentCGF}
\title{CGF Object for the sum of independent random variables}
\usage{
sumOfIndependentCGF(cgf_list, block_size = NULL, iidReps = NULL, ...)
}
\arguments{
\item{cgf_list}{A non-empty list of CGF objects. Each of class \code{"CGF"}.}

\item{block_size}{Either \code{NULL} or a positive integer specifying the block size for replication.
Default is \code{NULL}.}

\item{iidReps}{Either \code{NULL} or a positive integer specifying how many i.i.d. blocks
to expect. Default is \code{NULL}.}

\item{...}{Additional named arguments passed to \code{\link{createCGF}}, the \code{CGF} object creation function.}
}
\value{
A new \code{CGF} object.
}
\description{
Constructs a new CGF object representing the sum of multiple independent random variables.
}
\seealso{
\code{\link{adaptCGF}} for an example which also adapts the CGF object.
}
