% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{sumOfIndependentCGF}
\alias{sumOfIndependentCGF}
\title{CGF of a sum of independent random variables.}
\usage{
sumOfIndependentCGF(baseCGF_list)
}
\arguments{
\item{baseCGF_list}{List of an arbitrary number of CGF objects, each corresponding to an independent random variable.}
}
\value{
A CGF object representing the sum of independent random variables.
}
\description{
Evaluates the CGF object for the sum of independent(identical or non-identical) random variables.
}
