% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGF-operations.R
\name{sumOfIndependentCGF}
\alias{sumOfIndependentCGF}
\title{Sum of Independent CGFs}
\usage{
sumOfIndependentCGF(...)
}
\arguments{
\item{...}{CGF objects: Any number of CGF objects.
Each argument passed through \code{...} must be an object of class 'CGF', each corresponding to an independent random variable.}
}
\value{
A CGF object for the sum of independent random variables.
}
\description{
Computes the CGF object for the sum of independent (identical or non-identical) random variables.
}
